==> R/02-washb-geopair-relative-efficiency.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "Estimates of relative efficiency"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: pygments
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Relative efficacy based on non-parametric estimator from Imai, King, and Nall (2009)

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

# detach raster and gdistance packages because
# of conflicts with dplyr::select() and not
# needed
detach("package:gdistance", unload = TRUE)
detach("package:raster", unload = TRUE)


#----------------------------
# source the base functions
#----------------------------
source(here("R","washb-geopair-Functions.R"))

```



# Bangladesh

## Data prep

```{r load bangladesh data}
#-------------------------------
# load the formatted analysis
# data created by
# washb-geopair-data-processing.R
#-------------------------------

# anthropometry data
danth <- read_rds(here("data","bangl_analysis_anthro.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1)) 
  

# child development data
dchdev <- read_rds(here("data","bangl_analysis_chdev.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))

# diarrhea data
ddiar <- read_rds(here("data","bangl_analysis_diar.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))

# giardia data
dpara <- read_rds(here("data","bangl_analysis_parasite.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))

```

## Estimate ATE and relative efficacy
```{r estimate ATE and relative efficiency bangladesh}
#-------------------------------
# the estimate_MPCR_ATE function
# does not allow for missing values
# in the outcome, so subset each
# dataset to non-missing values
# before each estimation
#
# the estimate_MPCR_ATE function
# is in washb-geopair-Functions.R
#-------------------------------

#-------------------------------
# LAZ
#-------------------------------
d <- danth %>% filter(!is.na(laz))
ate_laz <- estimate_MPCR_ATE(Y=d$laz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# WAZ
#-------------------------------
d <- danth %>% filter(!is.na(waz))
ate_waz <- estimate_MPCR_ATE(Y=d$waz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# WHZ
#-------------------------------
d <- danth %>% filter(!is.na(whz))
ate_whz <- estimate_MPCR_ATE(Y=d$whz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# HCZ
#-------------------------------
d <- danth %>% filter(!is.na(hcz))
ate_hcz <- estimate_MPCR_ATE(Y=d$hcz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# EASQ Communication
#-------------------------------
d <- dchdev %>% filter(!is.na(z_easq_com))
ate_easq_com <- estimate_MPCR_ATE(Y=d$z_easq_com,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# EASQ Gross motor
#-------------------------------
d <- dchdev %>% filter(!is.na(z_easq_motor))
ate_easq_mot <- estimate_MPCR_ATE(Y=d$z_easq_motor,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# EASQ Personal-social
#-------------------------------
d <- dchdev %>% filter(!is.na(z_easq_pers))
ate_easq_ps <- estimate_MPCR_ATE(Y=d$z_easq_pers,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# CDI comprehension
#-------------------------------
d <- dchdev %>% filter(!is.na(z_cdi_comp))
ate_cdi_comp <- estimate_MPCR_ATE(Y=d$z_cdi_comp,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# CDI expression
#-------------------------------
d <- dchdev %>% filter(!is.na(z_cdi_expr))
ate_cdi_expr <- estimate_MPCR_ATE(Y=d$z_cdi_expr,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Diarrhea
#-------------------------------
d <- ddiar %>% filter(!is.na(diar7d))
ate_diar <- estimate_MPCR_ATE(Y=d$diar7d,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Giardia
#-------------------------------
d <- dpara %>% filter(!is.na(giar))
ate_giar <- estimate_MPCR_ATE(Y=d$giar,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Ascaris
#-------------------------------
d <- dpara %>% filter(!is.na(al))
ate_al <- estimate_MPCR_ATE(Y=d$al,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Trichuris
#-------------------------------
d <- dpara %>% filter(!is.na(tt))
ate_tt <- estimate_MPCR_ATE(Y=d$tt,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Hookworm
#-------------------------------
d <- dpara %>% filter(!is.na(hw))
ate_hw <- estimate_MPCR_ATE(Y=d$hw,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

```


```{r summarize relative efficiency bangladesh}
#-------------------------------
# Summarize relative efficacy
#-------------------------------
ate_list <- list(ate_laz,ate_waz,ate_whz,ate_hcz,
                 ate_easq_com,ate_easq_mot,ate_easq_ps,ate_cdi_comp,ate_cdi_expr,
                 ate_diar,ate_giar,ate_al,ate_tt,ate_hw)
out_list <- c("length-for-age z",
                         "weight-for-age z",
                         "weight-for-height z",
                         "head circumference z",
                         "EASQ communication z",
                         "EASQ gross motor z",
                         "EASQ personal-social z",
                         "CDI comprehension z",
                         "CDI expression z",
                         "Diarrhea",
                         "Giardia sp.",
                         "Ascaris sp.",
                         "Trichuris sp.",
                         "Hookworm")

ate_res <- foreach(esti = 1:length(ate_list), .combine = rbind) %do% {
  res <- data.frame(est = ate_list[[esti]]$est, var = ate_list[[esti]]$var,
                    est_min95 = ate_list[[esti]]$est_min95, est_max95 = ate_list[[esti]]$est_max95,
                    corr_w = ate_list[[esti]]$corr_w, corr_u = ate_list[[esti]]$corr_u,
                    eff_w = ate_list[[esti]]$eff_w, eff_u = ate_list[[esti]]$eff_u)
  res
}
ate_res$outcome_lab <- out_list
ate_res <- ate_res %>%
  mutate(outcome_lab = factor(outcome_lab, levels = out_list),
         country = "Bangladesh") %>%
  dplyr::select(country,outcome_lab, everything())

```


# Kenya

## Data prep

```{r kenya identify complete blocks}
#-------------------------------
# summarize the number of complete
# and incomplete blocks in the
# kenya trial based on the anthro
# data
#
# the code below identifies the
# 17 of 89 blocks that did not
# have a balanced design for
# Control and Nutrition clusters
# leaving 72 complete blocks of
# 4 clusters (2 control, 2 nutrition)
#-------------------------------
kcl <- read_rds(here("data","kenya_analysis_anthro.rds")) 

kcl2 <- kcl %>%
  dplyr::select(block,tr,clusterid) %>%
  group_by(block,tr,clusterid) %>%
  slice(1) %>%
  group_by(block,tr) %>%
  summarize(n=n(), .groups = "drop") %>%
  pivot_wider(id_cols = "block",names_from="tr",values_from="n") %>%
  mutate(block_comp = rowSums(across(everything())),
         block_comp_c = ifelse(Control==2,1,0),
         block_comp_n = ifelse(`Nutrition` + `Nutrition + WSH` == 2,1,0)
           )

table(is.na(kcl2$block_comp))
table(kcl2$block_comp_c,kcl2$block_comp_n, useNA = "ifany")

k_incomplete <- kcl2 %>%
  filter(block_comp_c==0 | is.na(block_comp_c) | block_comp_n==0 | is.na(block_comp_n) )

k_incomplete$block
```


```{r load Kenya data}
#-------------------------------
# load the formatted analysis
# data created by
# washb-geopair-data-processing.R
#
# in each dataset, exclude 
# the 10 matched blocks that do
# not have treatment contrasts
# for nutrition vs. control
#-------------------------------

# anthropometry data
dkanth <- read_rds(here("data","kenya_analysis_anthro.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control","Passive Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))


# diarrhea data
dkdiar <- read_rds(here("data","kenya_analysis_diar.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

# parasite data
dkpara <- read_rds(here("data","kenya_analysis_parasite.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

# child development data
dkchdev <- read_rds(here("data","kenya_analysis_chdev.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

```



## Estimate ATE and relative efficiency
```{r estimate ATE and relative efficiency kenya}
#-------------------------------
# the estimate_MPCR_ATE function
# does not allow for missing values
# in the outcome, so subset each
# dataset to non-missing values
# before each estimation
#-------------------------------

#-------------------------------
# LAZ
#-------------------------------
d <- dkanth %>% filter(!is.na(laz)) 
k_ate_laz <- estimate_MPCR_ATE(Y=d$laz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# WAZ
#-------------------------------
d <- dkanth %>% filter(!is.na(waz)) 
k_ate_waz <- estimate_MPCR_ATE(Y=d$waz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# WHZ
#-------------------------------
d <- dkanth %>% filter(!is.na(whz))
k_ate_whz <- estimate_MPCR_ATE(Y=d$whz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# HCZ
#-------------------------------
d <- dkanth %>% filter(!is.na(hcz))
k_ate_hcz <- estimate_MPCR_ATE(Y=d$hcz,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# EASQ Communication
#-------------------------------
d <- dkchdev %>% filter(!is.na(z_easq_com))
k_ate_easq_com <- estimate_MPCR_ATE(Y=d$z_easq_com,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# EASQ Gross motor
#-------------------------------
d <- dkchdev %>% filter(!is.na(z_easq_motor))
k_ate_easq_mot <- estimate_MPCR_ATE(Y=d$z_easq_motor,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# EASQ Personal-social
#-------------------------------
d <- dkchdev %>% filter(!is.na(z_easq_pers))
k_ate_easq_ps <- estimate_MPCR_ATE(Y=d$z_easq_pers,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)


#-------------------------------
# Diarrhea
#-------------------------------
d <- dkdiar %>% filter(!is.na(diar7d))
k_ate_diar <- estimate_MPCR_ATE(Y=d$diar7d,Z=d$tr01,grp=d$clusterid,match=d$block,data=d)

#-------------------------------
# Giardia
#-------------------------------
d <- dkpara %>% filter(!is.na(giar))
k_ate_giar <- estimate_MPCR_ATE(Y=d$giar,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# Ascaris
#-------------------------------
d <- dkpara %>% filter(!is.na(al))
k_ate_al <- estimate_MPCR_ATE(Y=d$al,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# Trichuris
#-------------------------------
d <- dkpara %>% filter(!is.na(tt))
k_ate_tt <- estimate_MPCR_ATE(Y=d$tt,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# Hookworm
#-------------------------------
d <- dkpara %>% filter(!is.na(hw))
k_ate_hw <- estimate_MPCR_ATE(Y=d$hw,Z=d$tr01,grp=d$clusteridr2,match=d$block,data=d)

#-------------------------------
# detach experiment and spaMM packages
# because they uses MASS, which
# conflicts with dplyr::select()
#-------------------------------
detach("package:experiment", unload = TRUE)
detach("package:spaMM", unload = TRUE)
detach("package:MASS", unload = TRUE)


```


```{r summarize relative efficiency kenya}
#-------------------------------
# Summarize relative efficiency
#-------------------------------
k_ate_list <- list(k_ate_laz,k_ate_waz,k_ate_whz,k_ate_hcz,
                 k_ate_easq_com,k_ate_easq_mot,k_ate_easq_ps,
                 k_ate_diar,k_ate_giar,k_ate_al,k_ate_tt,k_ate_hw)
k_out_list <- c("length-for-age z",
                         "weight-for-age z",
                         "weight-for-height z",
                         "head circumference z",
                         "EASQ communication z",
                         "EASQ gross motor z",
                         "EASQ personal-social z",
                         "Diarrhea",
                         "Giardia sp.",
                         "Ascaris sp.",
                         "Trichuris sp.",
                         "Hookworm")

k_ate_res <- foreach(esti = 1:length(k_ate_list), .combine = rbind) %do% {
  res <- data.frame(est = k_ate_list[[esti]]$est, var = k_ate_list[[esti]]$var,
                    est_min95 = k_ate_list[[esti]]$est_min95, est_max95 = k_ate_list[[esti]]$est_max95,
                    corr_w = k_ate_list[[esti]]$corr_w, corr_u = k_ate_list[[esti]]$corr_u,
                    eff_w = k_ate_list[[esti]]$eff_w, eff_u = k_ate_list[[esti]]$eff_u)
  res
}
k_ate_res$outcome_lab <- k_out_list
k_ate_res <- k_ate_res %>%
  mutate(outcome_lab = factor(outcome_lab, levels = out_list),
         country = "Kenya") %>%
  dplyr::select(country,outcome_lab, everything())

```
# Save pair-level summary dataset

Save a dataset that includes pair-level means and treatment effects. These estimates will be used in later analysis and plotting

```{r assemble pair level data}
#-------------------------------
# get the pair-level estimates
# Bangladesh
#-------------------------------
dcl_b <- foreach(i = 1:length(ate_list), .combine = rbind) %do% {
  dcli <- ate_list[[i]]$dcl
  dcli$block <- row.names(dcli)
  dcli$outcome_lab <- out_list[i]
  dcli$country = "Bangladesh"
  dcli
}
#-------------------------------
# get the pair-level estimates
# Kenya
#-------------------------------
dcl_k <- foreach(i = 1:length(k_ate_list), .combine = rbind) %do% {
  dcli <- k_ate_list[[i]]$dcl
  dcli$block <- row.names(dcli)
  dcli$outcome_lab <- k_out_list[i]
  dcli$country = "Kenya"
  dcli
}

#-------------------------------
# stack the data and convert
# some of the variables to 
# factors
#-------------------------------
dcl <- bind_rows(dcl_b,dcl_k) %>%
  mutate(outcome_lab=factor(outcome_lab, levels= out_list),
         country = factor(country)) %>%
  select(country,outcome_lab,block,everything()) %>%
  arrange(country,outcome_lab,block)

write_rds(dcl, file = here("output","washb_pair_level_summary_estimates.rds"))
write_csv(dcl, file = here("output","washb_pair_level_summary_estimates.csv"))


```

# Estimate 95% CIs for relative efficiency

Use a bootstrap resampling approach, resampling pairs with replacement, to estimate the 95% confidence interval for relative efficacy of geographic pair matching with 90 matched pairs. This will enable us to show whether lower efficacy at smaller sample sizes is significantly lower, and not just a function of sampling variability.

```{r bangladesh bootstrap CIs for relative efficiency}
#-------------------------------
# pivot wider to make the bootstrap
# easy
#-------------------------------
dboot <- dcl %>%
  filter(country == "Bangladesh") %>%
  group_by(outcome_lab)

# estimate outcome correlation, weighted by pair-wise N
corx <- function(x,y,w) {
    xw <- x*w
    yw <- y*w
    wcorr <- 2*cov(xw,yw) / ( var(xw) + var(yw) )
    return(wcorr)
}

#-------------------------------
# boostrap resample within
# each outcome group
# 1000 iterations
#-------------------------------
cor_boot_b <- foreach(iteri = 1:1000, .combine = rbind) %dopar% {
  set.seed(iteri)
  di <- dboot %>% 
    slice_sample(prop = 1, replace = TRUE)
  corest <- di %>% group_map( ~ corx(x=.x$Y1bar, y=.x$Y0bar,w = .x$w))
  
  res <- data.frame(iter = iteri, outcome_lab = levels(di$outcome_lab), corr = unlist(corest))
  res
}

cor_boot_b2 <- cor_boot_b %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl$outcome_lab)))

#-------------------------------
# calculate percentile 95% CIs 
# from the bootstrap distribution
#-------------------------------
cor_boot_ci_b <- cor_boot_b2 %>%
  group_by(outcome_lab) %>%
  mutate(cor_min95 = quantile(corr, probs = 0.025),
         cor_max95 = quantile(corr, probs = 0.975)) %>%
  slice(1) %>%
  select(outcome_lab, cor_min95, cor_max95)

#-------------------------------
# transform correlation 95% CIs 
# into relative efficacy
#-------------------------------
releff_ci_b <- cor_boot_ci_b %>%
  mutate(releff_min95 = 1/(1-cor_min95),
         releff_max95 = 1/(1-cor_max95)) %>%
  mutate(country = "Bangladesh")

#-------------------------------
# join to the existing estimates
# of relative efficiency
#-------------------------------
ate_res2 <- ate_res %>%
  left_join(releff_ci_b, by=c("country","outcome_lab"))

```

```{r kenya bootstrap CIs for relative efficiency}
#-------------------------------
# pivot wider to make the bootstrap
# easy
#-------------------------------
dboot <- dcl %>%
  filter(country == "Kenya") %>%
  mutate(outcome_lab = droplevels(outcome_lab)) %>%
  group_by(outcome_lab) 

#-------------------------------
# boostrap resample within
# each outcome group
# 1000 iterations
#-------------------------------
cor_boot_k <- foreach(iteri = 1:1000, .combine = rbind) %dopar% {
  set.seed(iteri)
  di <- dboot %>% 
    slice_sample(prop = 1, replace = TRUE)
  corest <- di %>% group_map( ~ corx(x=.x$Y1bar, y=.x$Y0bar,w = .x$w))
  
  res <- data.frame(iter = iteri, outcome_lab = levels(di$outcome_lab), corr = unlist(corest))
  res
}

cor_boot_k2 <- cor_boot_k %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl$outcome_lab)))

#-------------------------------
# calculate percentile 95% CIs 
# from the bootstrap distribution
#-------------------------------
cor_boot_ci_k <- cor_boot_k2 %>%
  group_by(outcome_lab) %>%
  mutate(cor_min95 = quantile(corr, probs = 0.025),
         cor_max95 = quantile(corr, probs = 0.975)) %>%
  slice(1) %>%
  select(outcome_lab, cor_min95, cor_max95)

#-------------------------------
# transform correlation 95% CIs 
# into relative efficiency
#-------------------------------
releff_ci_k <- cor_boot_ci_k %>%
  mutate(releff_min95 = 1/(1-cor_min95),
         releff_max95 = 1/(1-cor_max95)) %>%
  mutate(country = "Kenya")

#-------------------------------
# join to the existing estimates
# of relative efficiency
#-------------------------------
k_ate_res2 <- k_ate_res %>%
  left_join(releff_ci_k, by=c("country","outcome_lab"))

```


# Save ATE and Relative Efficiency Estimates
```{r save ate and reff estimates}
#-------------------------------
# Estimate control and intervention
# means
#-------------------------------
dcl_means <- dcl %>%
  group_by(country,outcome_lab) %>%
  summarize(Y0bar = weighted.mean(Y0bar,w = n0),
            Y1bar = weighted.mean(Y1bar,w = n1),
            .groups = "keep"
            )


ate_res3 <- ate_res2 %>%
  bind_rows(k_ate_res2) %>%
  left_join(dcl_means, by = c("country","outcome_lab")) %>%
  select(country, outcome_lab, Y0bar, Y1bar, everything())

write_rds(ate_res3, file = here("output","washb_ATE_releff_estimates.rds"))
write_csv(ate_res3, file = here("output","washb_ATE_releff_estimates.csv"))

```


# Session Info
```{r session info}
sessionInfo()
```
