==> R/03-washb-geopair-estimator-comparison.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "Estimate empirical gains in efficiency from pair matching and adjusted analyses (vs unmatched estimator)"
author: "Christine Tedijanto christine.tedijanto@ucsf.edu and Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

In this analysis, we compare the estimated relative efficiency based on analytic methods with the empirical relative efficiency. The empirical relative efficiency is calculated by comparing the variance of the treatment effect in an analysis that does not account for matching to the variance of the treatment effect using the Imai matched pair estimator. An additional analysis controls for subdistricts (zillas in Bangladesh, sub-counties in Kenya) and compares the relative efficiency of that estimator to an unmatched analysis. 

# Preamble

```{r preamble, message = FALSE}
library(here)
here()

library(cowplot) # for plot_grid

#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

#----------------------------
# source the base functions
#----------------------------
source(here("R","washb-geopair-Functions.R"))

```

# Data prep

```{r load gps data}
#-------------------------------
# load the formatted gps
# data created by
# washb-geopair-gps-data-processing.R
#
# note: only includes GADM (not GPS) data
#-------------------------------

## Bangladesh -----
dbgps <- read_rds(here("data","bangl_analysis_gadm.rds")) %>%
  mutate(block = as.character(block),
         blockf = factor(block))

## Kenya -----
dkgps <- read_rds(here("data","kenya_analysis_gadm.rds")) %>%
  mutate(block = as.character(block),
         blockf = factor(block))
```

```{r load cluster level means and relative efficiency}
#-------------------------------
# read-in cluster level means and relative efficiency
# these were estimated in the file
# washb-geopair-relative-efficiency.Rmd
#-------------------------------
dcl <- read_rds(here("output", "washb_pair_level_summary_estimates.rds")) 

ate_res <- read_rds(here("output", "washb_ATE_releff_estimates.rds"))
```

```{r outcome labels}
outcome_lab_key <- data.frame(outcome = c("laz", "waz", "whz", "hcz",
                                          "z_easq_com", "z_easq_motor", "z_easq_pers", "z_cdi_comp", "z_cdi_expr",
                                          "diar7d", "giar", "al", "tt", "hw"),
                              outcome_lab = c("length-for-age z", "weight-for-age z", "weight-for-height z", "head circumference z",
                                              "EASQ communication z", "EASQ gross motor z", "EASQ personal-social z",
                                              "CDI comprehension z", "CDI expression z",
                                              "Diarrhea", "Giardia sp.", "Ascaris sp.", "Trichuris sp.", "Hookworm"),
                              outcome_group = c(rep("Child growth", 4),
                                                 rep("Child development", 5),
                                                 rep("Infectious disease", 5)))
```

## Create subdistrict dataframe

```{r subdistrict data}
# blocks are not all nested within subdistricts (15 / 347 instances with repeats)
dsubdist <- dbgps %>% 
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr == "Control", 0, 1)) %>% 
  dplyr::select(block, subdist = ADM3_PCODE, tr01) %>% # use lowest admin level
  distinct() %>% 
  mutate(country = "Bangladesh") %>% 
  bind_rows(dkgps %>%
              filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
              mutate(tr01 = ifelse(tr == "Control", 0, 1)) %>% 
              dplyr::select(block, subdist = ADM2_PCODE, tr01) %>% # use lowest admin level
              distinct() %>% 
              mutate(country = "Kenya")) %>% 
  group_by(country, block, tr01) %>% 
  # select first subdistrict for each block
  slice(1) %>% 
  ungroup()
```

# Estimate ATEs 

Consider 3 estimators: pair-matched, subdistrict-adjusted, and unadjusted (unmatched)

```{r run regression models}
# create shorter cluster-level means dataset with variables of interest
dcl_2 <- dcl %>%
  remove_rownames() %>% 
  dplyr::select(country, outcome_lab, block, Y1bar, Y0bar, n1,n0)

## `get_reg_results`: function to run matched, unmatched, and subdistrict regressions -----
## inputs:
# - c (character): current country of interest
# - o (character): outcome for regression models
## outputs:
# - dataframe with country, outcome, and estimated ATE and
# variance of ATE from unmatched, matched and subdistrict models
get_reg_results <- function(c, o){
  temp_df <- dcl_2 %>% 
    filter(country == c, outcome_lab == o) 
  
  temp_dfl <- temp_df %>% 
    pivot_longer(cols = c(Y1bar, Y0bar),
                 names_to = "name", 
                 values_to = "Ybar") %>% 
    mutate(tr01 = ifelse(name=="Y1bar",1,0),
           n = ifelse(name=="Y1bar",n1,n0)) %>% 
    left_join(dsubdist, by = c("country", "block", "tr01"))
  
  # unmatched, non-parametric estimator
  nomatch <- estimate_UMCR_ATE(Y0bar=temp_df$Y0bar,
                               Y1bar=temp_df$Y1bar,
                               n0 = temp_df$n0,
                               n1 = temp_df$n1)
  ate_nomatch <- nomatch$ate_est
  var_nomatch <- nomatch$ate_var
  
  # matched regression (analogous to a weighted paired t-test)
  # this is an internal validity check to make sure it is
  # similar to the non-parametric estimator, so that the
  # next subdistrict adjusted model is comparable, with just
  # a coarser level of stratification
  matchreg <- lm(Ybar ~ tr01 + as.factor(block),
                  weights = n,
                  data = temp_dfl)
  ate_matchreg <- summary(matchreg)$coefficients["tr01", "Estimate"]
  var_matchreg <- summary(matchreg)$coefficients["tr01", "Std. Error"]^2
  
  # regression adjusting for subdistrict
  subdreg <- lm(Ybar ~ tr01 + as.factor(subdist),
                 weights = n,
                 data = temp_dfl)
  ate_subdreg <- summary(subdreg)$coefficients["tr01", "Estimate"]
  var_subdreg <- summary(subdreg)$coefficients["tr01", "Std. Error"]^2
  
  ret <- data.frame(country = c,
                    outcome_lab = o,
                    ate_matchreg = ate_matchreg,
                    var_matchreg = var_matchreg,
                    ate_nomatch = ate_nomatch,
                    var_nomatch = var_nomatch,
                    ate_subdreg = ate_subdreg,
                    var_subdreg = var_subdreg)
  
  return(ret)
}

## run apply to get regression results -----

# create unique list of country-outcome pairs for apply
country_outcome_uniq <- dcl_2 %>% distinct(country, outcome_lab)

reg_results <- mapply(get_reg_results,
                      c = country_outcome_uniq$country,
                      o = country_outcome_uniq$outcome_lab,
                      SIMPLIFY = FALSE) %>% 
  bind_rows()

all_results <- reg_results %>% 
  left_join(ate_res %>% 
              dplyr::select(country, outcome_lab, ate_mpcr = est, var_mpcr = var,
                            eff_w, eff_u),
            by = c("country", "outcome_lab")) %>% 
  left_join(outcome_lab_key, by = "outcome_lab") %>%
  # order vars
  dplyr::select(country,outcome_group,outcome,outcome_lab,eff_w,eff_u,everything()) %>%
  # calculate the relative efficiency for the matched pair estimator and the 
  # subdistrict adjusted estimator
  mutate(reff_mpcr = var_nomatch / var_mpcr,
         reff_subdreg = var_nomatch / var_subdreg)

```

# Visualize results

First, just a quick internal validity check to ensure the weighted least squares regression that conditions on matched pairs, equivalent to a weighted t-test, yields identical estimates of the ATE and very close estimates of the variance. The variance won't be identical but it should be very, very close.

```{r compare ate for wls and mpcr, fig.width=6}
plot_ate_comp <- ggplot(data=all_results, aes(x=ate_mpcr,y=ate_matchreg)) +
  facet_grid(.~country) + 
  geom_abline(intercept=0,slope=1, color="gray40") +
  geom_point() + 
  labs(x="non-parametric pair-matched",
       y="weighted t-test",
       title = "Comparison of ATE estimates across outcomes"
       ) +
  theme_minimal()
  
plot_ate_comp


```

```{r compare variance for wls and mpcr, fig.width=6}
plot_var_comp <- ggplot(data=all_results, aes(x=var_mpcr,y=var_matchreg)) +
  facet_grid(.~country) + 
  geom_abline(intercept=0,slope=1, color="gray40") +
  geom_point() + 
  labs(x="non-parametric pair-matched",
       y="weighted t-test",
       title = "Comparison of variance estimates across outcomes"
       ) +
  theme_minimal()
  
plot_var_comp


```

Outcomes are arranged in descending order of relative efficiency of the non-parametric estimator compared to the unmatched estimator.

```{r visualize results}
outcome_colors <- c("Child development" = "#D9B86A",
                    "Child growth" = "#9B9F7C",
                    "Infectious disease" = "#5986B0")

# determine ordering for outcomes
# sort by relative efficiency observed for MPCR estimator
orderb <- all_results %>%
  filter(country == "Bangladesh") %>% 
  arrange(desc(reff_mpcr)) %>%
  pull(outcome_lab)

orderk <- all_results %>%
  filter(country == "Kenya") %>% 
  arrange(desc(reff_mpcr)) %>%
  pull(outcome_lab)

```

## Bangladesh
```{r bangladesh fig, fig.width = 5, fig.height = 3.2}
all_results %>%
  filter(country == "Bangladesh") %>% 
  pivot_longer(cols = c("var_matchreg", "var_mpcr", "var_subdreg"),
               names_pattern = "var_(.*)",
               names_to = "model",
               values_to = "var") %>%
  filter(model != "matchreg") %>% # remove pair-matched regression (same as mpcr, above)
  mutate(rel_eff = var_nomatch/var) %>% # relative efficiency is ratio, comparing variance of unmatched analysis to variance from other analyses
  mutate(outcome_lab = factor(outcome_lab, levels = orderb)) %>% 
  ggplot(aes(x = outcome_lab, y = rel_eff)) + 
  geom_hline(yintercept=1, linetype = "dashed") +
  geom_point(aes(fill = outcome_group, shape = model),
             size = 2, alpha = 0.7) +
  scale_shape_manual(values = c("mpcr" = 24, "subdreg" = 21),
                     labels = c("mpcr" = "pair-matched", "subdreg" = "subdistrict adj")) +
  scale_fill_manual(values = outcome_colors) +
  labs(x = NULL, y = "rel efficiency compared to unmatched analysis") +
  coord_flip(ylim = c(0.8,3.6)) +
  guides(fill = "none") +
  theme_bw() +
  theme(
    legend.position = "bottom",
        )
```

## Kenya

```{r kenya fig, fig.width = 5, fig.height = 3.2}
all_results %>%
  filter(country == "Kenya") %>% 
  pivot_longer(cols = c("var_matchreg", "var_mpcr", "var_subdreg"),
               names_pattern = "var_(.*)",
               names_to = "model",
               values_to = "var") %>%
  filter(model != "matchreg") %>% # remove pair-matched regression (same as mpcr, above)
  mutate(rel_eff = var_nomatch/var) %>% # relative efficiency is ratio, comparing variance of unmatched analysis to variance from other analyses
  mutate(outcome_lab = factor(outcome_lab, levels = orderk)) %>% 
  ggplot(aes(x = outcome_lab, y = rel_eff)) + 
  geom_hline(yintercept=1, linetype = "dashed") +
  geom_point(aes(fill = outcome_group, shape = model),
             size = 2, alpha = 0.7) +
  scale_shape_manual(values = c("mpcr" = 24, "subdreg" = 21),
                     labels = c("mpcr" = "pair-matched", "subdreg" = "subdistrict adj")) +
  scale_fill_manual(values = outcome_colors) +
  labs(x = NULL, y = "rel efficiency compared to unmatched analysis") +
  coord_flip(ylim = c(0.8,3.6)) +
  guides(fill = "none") +
  theme_bw() +
  theme(
    legend.position = "bottom",
        )
```

## Additional comparison figures

```{r compare, fig.width = 6, fig.height = 7}
wt_eff_fig <- all_results %>% 
  mutate(rel_eff = var_nomatch / var_mpcr) %>% 
  ggplot(aes(x = eff_w, y = rel_eff, fill = outcome_group)) +
  geom_abline(intercept=0,slope=1) +
  geom_point(pch = 21, size = 2, alpha = 0.8) +
  scale_fill_manual(values = outcome_colors) +
  coord_cartesian(xlim=c(1,3.6),ylim=c(1,3.6))+
  facet_wrap(.~country) +
  labs(x = "relative efficiency predicted by weighted correlation",
       y = "relative efficiency observed",
       fill = "outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 8))

unwt_eff_fig <- all_results %>% 
  mutate(rel_eff = var_nomatch / var_mpcr) %>% 
  ggplot(aes(x = eff_u, y = rel_eff, fill = outcome_group)) +
  geom_abline(intercept=0,slope=1) +
  geom_point(pch = 21, size = 2, alpha = 0.8) +
  scale_fill_manual(values = outcome_colors) +
  coord_cartesian(xlim=c(1,3.6),ylim=c(1,3.6))+
  facet_wrap(.~country) +
  labs(x = "relative efficiency predicted by unweighted correlation",
       y = "relative efficiency observed",
       fill = "outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 8))

plot_grid(wt_eff_fig,
          unwt_eff_fig,
          nrow = 2)
```

# Save results

```{r save results}
write_csv(all_results,file=here("output","washb_geopair_estimator_comparison.csv"))
write_rds(all_results,file=here("output","washb_geopair_estimator_comparison.rds"))


```

# Session Info

```{r session info}
sessionInfo()
```
