==> R/06-washb-geopair-relative-efficiency-figures.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "Make relative efficiency summary figures"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: pygments
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

This script relies on output files created by upstream analysis files to make summary figures. The upstream analysis files include: `02-washb-geopair-relative-efficiency.Rmd`, `03-washb-geopair-estimator-comparison.Rmd`, `04-washb-geopair-moransi.Rmd`, and `05-washb-geopair-estimate-icc.R`. 

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

# detach MASS and spaMM/experiment packages
# detach raster and gdistance packages 
# of conflicts with dplyr::select() and not
# needed
detach("package:experiment", unload = TRUE)
detach("package:spaMM", unload = TRUE)
detach("package:MASS", unload = TRUE)
detach("package:gdistance", unload = TRUE)
detach("package:raster", unload = TRUE)


#----------------------------
# source the base functions
#----------------------------
source(here("R","washb-geopair-Functions.R"))

```


# Load previous estimates
```{r load ate and reff estimates}
#-------------------------------
# load previous estimates of
# predicted relative efficiency
# based on pair-level correlation
#-------------------------------
ate_res <- read_rds(file = here("output","washb_ATE_releff_estimates.rds"))

#-------------------------------
# load empirical estimates of
# relative efficiency for pair matched
# and sub-district adjusted estimators
#-------------------------------
estcomp_res <- read_rds(file = here("output","washb_geopair_estimator_comparison.rds"))

#-------------------------------
# load empirical estimates of
# Moran's I for outcomes
#-------------------------------
morani_res <- read_rds(file = here("output","washb-geopair-morans-I-estimates.rds"))

#-------------------------------
# load empirical estimates of
# the ICC for outcomes
#-------------------------------
icc_res <- read_rds(file = here("output","washb-geopair-icc-estimates.rds"))
```


# Break even correlation

The power of the matched t-test for a two-tailed 0.05 test where the power of the unmatched test is held at 0.80 and the pairs are uncorrelated

```{r power of matched vs unmatched}

#------------------------------
# power of the matched
# compared to unmatched test
# r = 0
#------------------------------
ns <- 2:50
sdu <- 1
powr0 <- foreach(ni = ns, .combine = rbind) %do% {
  pu <- power.t.test(n=ni,sd=sdu, power=0.8, type="two.sample")
  pm <- power.t.test(n=ni,delta=pu$delta, sd=sqrt(2)*sdu, sig.level= 0.05, type = "one.sample")
  res <- data.frame(n=ni, powm = pm$power, powu = pu$power,sdm=pm$sd,sdu = sdu)
}

ggplot(data=powr0, aes(x=n)) +
  geom_hline(yintercept = 0.8, color = "gray40") + 
  geom_line(aes(y=powm)) +
  scale_y_continuous(breaks=seq(0,1,by=0.1))+
  coord_cartesian(ylim = c(0,1)) +
  labs(x = "number of pairs", y = "power") +
  theme_minimal() 
```

Solve for break even correlations of matched versus unmatched designs. This is an adaptation of Figure 4 in Martin et al. (1993)(https://pubmed.ncbi.nlm.nih.gov/8456215/), where they present a function of within-pair correlation, $r_{xy}$, whose squared value equals the within-pair correlation, ($r_{yy}$ in their notation). Below is the break-even correlation for within-pair outcomes.

```{r break even correlation}
#------------------------------
# solve for break-even r
#------------------------------
ns <- 3:50
sdu <- 1
breakrs <- foreach(ni = ns, .combine = rbind) %do% {
  pu <- power.t.test(n=ni,sd=sdu, power=0.8, sig.level= 0.05, type="two.sample")
  pm <- power.t.test(n=ni,delta=pu$delta, power=0.8,sd=NULL, sig.level= 0.05, type = "one.sample")
  ri = 1 - ( pm$sd^2/(2*(sdu)^2) )
  res <- data.frame(n=ni, powm = pm$power, powu = pu$power,sdm=pm$sd,sdu = sdu, r=ri)
}
ggplot(data=breakrs, aes(x=n,y=r)) +
  geom_line(aes(), color = "gray40") + 
  geom_point() +
  scale_y_continuous(breaks=seq(0,1,by=0.1))+
  labs(x = "number of pairs", y = "break-even correlation (r)") +
  annotate("text",x = 10, y=0.225, label = "Favors a matched design", hjust = 0)+
  annotate("text",x = 1, y=0.025, label = "Favors unmatched design", hjust = 0)+
  coord_cartesian(ylim = c(0,0.6), xlim=c(0,50)) +
  theme_minimal() 

#------------------------------
# solve for break even r
# by iteration
# (not used) 100% consistent
# with the analytic solution, above
# just did this to confirm
#------------------------------
# ns <- 2:50
# sdu <- 1
# powrs <- foreach(ni = ns, .combine = rbind) %do% {
#   pu <- power.t.test(n=ni,sd=sdu, power=0.8, type="two.sample")
#   ri <- 0
#   powm <- 0
#   while(ri < 1 & powm < 0.8) {
#     pm <- power.t.test(n=ni,delta=pu$delta, sd=sqrt(2)*sdu*sqrt(1-ri), sig.level= 0.05, type = "one.sample")
#     powm <- pm$power
#     ri <- ri+0.001
#   }
#   res <- data.frame(n=ni, powm = pm$power, powu = pu$power,sdm=pm$sd,sdu = sdu, r=ri)
# }


```

# General plot aesthetics

```{r general plot aesthetics}

# group outcomes for plot aesthetics
dreffplot <- ate_res 
dreffplot$outcome_group = c(rep("Child growth",4),rep("Child development",5),rep("Infectious disease",5) , rep("Child growth",4),rep("Child development",3),rep("Infectious disease",5))

# make shorter outcome labels
dreffplot$outcome_lab2 = c( c("LAZ","WAZ","WHZ","HCZ","EASQ Communication","EASQ Gross motor","EASQ Personal-social","CDI Comprehension","CDI Expression","Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm"), c("LAZ","WAZ","WHZ","HCZ","EASQ Communication","EASQ Gross motor","EASQ Personal-social","Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm"))

outcome_colors <- c("Child development" = "#D9B86A",
                    "Child growth" = "#9B9F7C",
                    "Infectious disease" = "#5986B0")

pcols <- c(cbpal[2],cbpal[4],"gray50")


```

# Correlation and relative efficiency function

```{r plot relative efficiency, warning = FALSE}

# data frame for function
d_reff <- data.frame(rho = seq(0,1,by=0.01)) %>%
  mutate(reff = 1/(1-rho))

# label nudging (tedious! bespoke! careful!)
dreffplot2 <- dreffplot %>%
  mutate(outcome_group = factor(outcome_group),
         country = factor(country)) %>%
  # tedious x and y nudge coordinates for labels to optimize display
  mutate(lab_xnudge = case_when(
           outcome_lab2 == "LAZ" & country == "Bangladesh" ~ 0.015,
           outcome_lab2 == "WAZ" & country == "Bangladesh" ~ 0.015,
           outcome_lab2 == "HCZ" & country == "Bangladesh" ~ 0.015,
           # outcome_lab2 == "Diarrhea" & country == "Bangladesh" ~ 0.015,
           outcome_lab2 == "Hookworm" & country == "Bangladesh" ~ 0.015,
           outcome_lab2 == "WHZ" & country == "Kenya" ~ 0.015,
           outcome_lab2 == "WAZ" & country == "Kenya" ~ 0.015,
           outcome_lab2 == "HCZ" & country == "Kenya" ~ 0.015,
           outcome_lab2 %in% c("EASQ Communication","EASQ Personal-social","CDI Expression") & country == "Bangladesh" ~ 0.015,
           TRUE ~ as.numeric(-0.015)
         ),
         lab_ynudge = case_when(
           outcome_lab2 == "Hookworm" & country == "Kenya" ~ 0.02,
           outcome_lab2 == "Trichuris sp." & country == "Kenya" ~ 0.02,
           outcome_lab2 == "WHZ" & country == "Kenya" ~ -0.02,
           outcome_lab2 == "Diarrhea" & country == "Bangladesh" ~ 0.01,
           outcome_lab2 == "EASQ Gross motor" & country == "Bangladesh" ~ 0.01,
           outcome_lab2 == "EASQ Personal-social" & country == "Bangladesh" ~ -0.02,
           outcome_lab2 == "EASQ Personal-social" & country == "Kenya" ~ 0.02,
           outcome_lab2 == "CDI Comprehension" & country == "Bangladesh" ~ 0.02,
           outcome_lab2 == "CDI Expression" & country == "Bangladesh" ~ -0.03,
           TRUE ~ as.numeric(0)
         ),
         lab_hjust = ifelse(lab_xnudge < 0 , 1, 0)
         )

# make plot
plot_reff <- ggplot(data=dreffplot2, aes(x = corr_w, y = eff_w)) +
  # facet over country
  facet_grid(~country) +
  # add function line
  geom_line(data=d_reff,aes(x=rho,y=reff),alpha = 0.4, linetype = "dashed", show.legend = FALSE) +
  # plot results
  # this offset is very idiosyncratic to optimize display -- fragile code!
  geom_text(aes(label = outcome_lab2, color = outcome_group), 
            hjust = dreffplot2$lab_hjust, 
            nudge_x = dreffplot2$lab_xnudge,
            nudge_y = dreffplot2$lab_ynudge,
            size = 2.5, show.legend = FALSE) +
  geom_point(aes(color = outcome_group), size = 2.2, alpha=1) + 
  geom_point(pch=21,bg=NA, size=2.2, show.legend = TRUE) +
  # refine aesthetics
  scale_color_manual(values = pcols, guide=guide_legend(title="Outcome group:")) +
  # scale_color_brewer(palette = "Dark2", guide=guide_legend(title="Outcome group:")) +
  scale_y_continuous(breaks=seq(1,3.2,by=0.2))+
  scale_x_continuous(breaks=seq(0,0.7,by=0.1)) +
  coord_cartesian(ylim=c(1,3.2),xlim = c(0,0.7)) +
  labs(x=expression(paste("Correlation between paired outcomes, (",italic(r),")")),y=expression(paste("Relative efficiency of geographic pair matching, (1-",italic(r),")")^-1), tag = "a") +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16)
  )
plot_reff

```

```{r plot relative efficiency transpose, warning = FALSE}

# data frame for function
d_reff <- data.frame(rho = seq(0,1,by=0.01)) %>%
  mutate(reff = 1/(1-rho))

# label nudging (tedious! bespoke! careful!)
xspacing <- 0.07
dreffplot2 <- dreffplot %>%
  mutate(outcome_group = factor(outcome_group),
         country = factor(country)) %>%
  # tedious x and y nudge coordinates for labels to optimize display
  mutate(
           lab_xnudge = case_when(
             outcome_lab %in% c("length-for-age z","weight-for-height z","head circumference z") ~ xspacing,
             outcome_lab %in%  c("weight-for-age z","Trichuris sp.","Hookworm") & country == "Kenya" ~ xspacing,
             outcome_group == "Child development" & country == "Bangladesh" ~ xspacing,
             outcome_group == "Diarrhea" & country == "Bangladesh" ~ -0.01,
             TRUE ~ as.numeric(-xspacing)
           ),
           lab_ynudge = case_when(
             outcome_lab %in% c("EASQ communication z","Giardia sp.") & country == "Bangladesh" ~ 0.01,
             outcome_lab == c("Ascaris sp.")  ~ 0.01,
             outcome_lab == "EASQ gross motor z" & country == "Bangladesh" ~ -0.01,
             outcome_lab %in% c("length-for-age z","Trichuris sp.") & country == "Kenya" ~ -0.02,
             TRUE ~ as.numeric(0)
           ),
         lab_hjust = ifelse(lab_xnudge < 0 , 1, 0)
         )

# make plot
# pcols <- c("#D9B86A", "#9B9F7C", "#5986B0")

plot_reff_t <- ggplot(data=dreffplot2, aes(x = eff_w, y = corr_w)) +
  # facet over country
  facet_grid(~country) +
  # add function line
  geom_line(data=d_reff,aes(y=rho,x=reff),alpha = 0.4, linetype = "dashed", show.legend = FALSE) +
  # plot results
  # this offset is very idiosyncratic to optimize display -- fragile code!
  geom_text(aes(label = outcome_lab, color = outcome_group),
            hjust = dreffplot2$lab_hjust,
            nudge_x = dreffplot2$lab_xnudge,
            nudge_y = dreffplot2$lab_ynudge,
            size = 2.5, show.legend = FALSE) +
  geom_point(aes(color = outcome_group), size = 2.2, alpha=1) + 
  geom_point(pch=21,bg=NA, size=2.2, show.legend = TRUE) +
  # refine aesthetics
  scale_color_manual(values = pcols, guide=guide_legend(title="outcome group:")) +
  # scale_color_brewer(palette = "Dark2", guide=guide_legend(title="outcome group:")) +
  scale_x_continuous(breaks=seq(1,3.4,by=0.4))+
  scale_y_continuous(breaks=seq(0,0.7,by=0.1)) +
  coord_cartesian(xlim=c(1,3.5),ylim = c(0,0.7)) +
  labs(y=expression(paste("Correlation between paired outcomes, (",italic(r),")")),x=expression(paste("Predicted relative efficiency of geographic pair matching, (1-",italic(r),")")^-1), tag = "a") +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor.y = element_blank(),
    plot.tag = element_text(face = "bold", size=16),
    strip.text = element_text(size=12)
  )
plot_reff_t

```

# Predicted vs observed relative efficiency

Compare the predicted (based on pair-level correlation) versus observed relative efficiency gains.

```{r predicted vs observed relative efficiency}
obs_v_pred_reff <- ggplot(data = estcomp_res, 
                          aes(x = eff_w, y = reff_mpcr, color = outcome_group)) +
  geom_abline(intercept=0,slope=1,color="gray70") +
  geom_point(pch = 19, size = 2.2, alpha = 0.8) +
  scale_color_manual(values = pcols, guide=guide_legend(title="Outcome group:")) +
  scale_x_continuous(breaks=seq(1,3.5,by=0.4))+
  scale_y_continuous(breaks=seq(1,3.5,by=0.4))+
  coord_cartesian(xlim=c(1,3.5),ylim=c(1,3.5))+
  facet_wrap(.~country) +
  labs(x = expression(paste("Predicted relative efficiency of geographic pair matching, (1-",italic(r),")")^-1),
       y = "Observed relative efficiency",
       tag = "b",
       fill = "Outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        # panel.grid.minor = element_blank(),
        plot.tag = element_text(face = "bold", size=16),
        strip.text = element_text(size=12)
        )

obs_v_pred_reff
```

Create a composite figure

```{r composite correlation figure, fig.height = 9}
plot_reff_composite <- grid.arrange(plot_reff_t,obs_v_pred_reff,ncol=1,nrow=2,heights=c(0.85,1))
ggsave(here("output","washb-geopair-reff-ests.pdf"),
       plot_reff_composite,
       device="pdf",
       height=180*1.14,
       width=180, 
       units = "mm")
```


# Correlation figure

Compare weighted versus unweighted correlation.  Weighted correlation is generally higher (consistent with results from Imai, King, Nall 2009), and relative efficiency estimates from weighted correlation (above) align far more closely with empirical estimates of relative efficiency.  For this reason, we use weighted estimates in the paper's primary comparisons.

```{r plot correlation}

# retrieve the break-even correlation for N = 10 pairs
breakr10 <- breakrs %>% 
  filter(n==10) %>%
  pull(r)

#----------------------------------
# create an unweighted vs weighted correlation plot
# pcols <- c("#D9B86A", "#9B9F7C", "#5986B0")
plot_corr <- ggplot(data = dreffplot, aes(x = corr_w, y = corr_u, color = outcome_group)) +
  facet_grid(~country) +
  geom_vline(xintercept=breakr10,color = "gray60", lwd=0.5, lty="dashed") +
  # geom_hline(yintercept=breakr10,color = "gray60", lwd=0.5, lty="dashed") + 
  annotate("text",x=breakr10+0.02,y=0.9,label="Break-even correlation\nfor n=10 pairs",hjust=0,size=3) +
  geom_abline(intercept = 0, slope = 1, color="gray70") +
  geom_point(pch=19, bg=NA, size=2.2, show.legend = TRUE) +
  geom_text(data = dreffplot %>% filter(corr_w - corr_u >0.2), 
            aes(label = outcome_lab, color = outcome_group),
            size = 2.5, hjust=0, nudge_x = 0.03, nudge_y=0.01,show.legend = FALSE) +
  scale_color_manual(values = pcols, guide=guide_legend(title="outcome group:")) +
  # scale_fill_manual(values = pcols) +
  # scale_color_brewer(palette = "Dark2", guide=guide_legend(title="outcome group:")) +
  scale_x_continuous(breaks=seq(-0.2,1,by=0.2))+
  scale_y_continuous(breaks=seq(-0.2,1,by=0.2))+
  labs(x = "Weighted correlation", y = "Unweighted correlation", tag = "a") +
  coord_cartesian(xlim=c(-0.2,1),ylim = c(-0.2,1)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),    
    plot.tag = element_text(face = "bold", size=16),
    strip.text = element_text(size=12)
  )


plot_corr

```

Compare the predicted (based on pair-level unweighted correlation) versus observed relative efficiency gains.

```{r plot obs releff vs weighted releff}


d_obs_v_pred_reff <- estcomp_res %>% 
  mutate(rel_eff = var_nomatch / var_mpcr)

# identify and label big discrepancies, > 0.4
obs_v_pred_wreff_labs <- d_obs_v_pred_reff %>%
  filter(abs(eff_u - rel_eff) >0.4) %>%
  mutate(plot_nudge_x = ifelse(outcome_lab %in% c("Ascaris sp.","Giardia sp."),-0.05,0.05),
         plot_nudge_x = ifelse(country=="Kenya" & outcome_lab %in% c("length-for-age z"),-0.05,plot_nudge_x),
         plot_hjust = ifelse( outcome_lab %in% c("Ascaris sp.","Giardia sp."),1,0),
         plot_hjust = ifelse( country=="Kenya" & outcome_lab %in% c("length-for-age z"),1,plot_hjust)
         )

  
obs_v_pred_wreff <- 
  ggplot(data=d_obs_v_pred_reff, aes(x = rel_eff, y = eff_u,color = outcome_group)) +
  geom_abline(intercept=0,slope=1,color="gray70") +
  geom_point(aes(),size=2.5, alpha=0.9) +
  geom_point(pch=21, bg=NA, size=2.5, show.legend = TRUE) +
  geom_text(data = obs_v_pred_wreff_labs, 
            aes(label = outcome_lab),
            size = 2.5, 
            hjust= obs_v_pred_wreff_labs$plot_hjust,
            nudge_y = obs_v_pred_wreff_labs$plot_nudge_y,
            nudge_x = obs_v_pred_wreff_labs$plot_nudge_x,
            show.legend = FALSE) +
  # scale_color_brewer(palette = "Dark2",guide=guide_legend(title="outcome group:")) +
  scale_color_manual(values = pcols, guide=guide_legend(title="Outcome group:")) +
  scale_x_continuous(breaks=seq(1,3.4,by=0.4))+
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  coord_cartesian(xlim=c(0.8,3.5),ylim=c(0.8,3.5))+
  facet_wrap(.~country) +
  labs(y = "Predicted relative efficiency, unweighted correlation",
       x = "Observed relative efficiency",
       tag = "b",
       fill = "Outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        panel.grid.minor = element_blank(),
        plot.tag = element_text(face = "bold", size=16),
        strip.text = element_text(size=12)
        )

obs_v_pred_wreff
```




```{r composite correlation plot, fig.height = 9}
plot_wgt_corr_composite <- grid.arrange(plot_corr,obs_v_pred_wreff,ncol=1,nrow=2,heights=c(0.85,1))
ggsave(here("output","washb-geopair-unweight-v-weight-corr.pdf"),
       plot_wgt_corr_composite,
       device="pdf",
       height=180*1.14,
       width=180,
       units = "mm")
```



# Relative efficiency by Moran's I, ICC, and prevalence

Plot the observed relative efficiency versus Moran's I, block-level ICC, and outcome prevalence (for binary outcomes only). Moran's I, ICC, and outcome prevalence were estimated using the control clusters only (excluding intervention clusters). It was possible to estimate the cluster-level ICC since we had repeated control clusters within each matched pair. 

```{r join icc and morani to reff ests}
#--------------------------------
# join empirical estimates of relative efficiency
# with ICC and Moran's I estimates
#--------------------------------
d_reff_icc <- estcomp_res %>%
  left_join(icc_res %>% rename(outcome_lab = outcome), 
            by = c("country","outcome_lab")) 

d_reff_morani <- estcomp_res %>%
  left_join(morani_res %>% dplyr::select(country,outcome_lab,starts_with("morani_")), 
            by = c("country","outcome_lab")) 

d_reff_prev <- estcomp_res %>%
  left_join(ate_res %>% dplyr::select(country,outcome_lab,Y0bar),
            by = c("country","outcome_lab")) 
  

```


```{r plot reff vs icc, warning = FALSE, message = FALSE}
plot_reff_icc <- ggplot(data = d_reff_icc, aes(x = icc, y = reff_mpcr)) +
  geom_smooth(method = "glm", se=FALSE, color="gray80") +
  geom_point(aes(color = outcome_group), pch = 19, size = 2.2, alpha = 0.8) +
  scale_color_manual(values = pcols, guide=guide_legend(title="Outcome group:")) +
  # scale_fill_brewer(palette = "Dark2", guide=guide_legend(title="outcome group:")) +
  scale_x_continuous(breaks=seq(0,0.12,by=0.04))+
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  coord_cartesian(xlim=c(0,0.12),ylim=c(1,3.5))+
  facet_grid(country~.) +
  labs(x = "Intra-cluster correlation (ICC)",
       y = "Observed relative efficiency",
       tag = "a",
       fill = "Outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_text(size = 11),
        legend.text = element_text(size = 11),
        # panel.grid.minor.y = element_blank(),
        plot.tag = element_text(face = "bold", size=16),
        strip.text.y = element_blank()
        )

plot_reff_icc

```

```{r plot reff vs morani, warning = FALSE, message = FALSE}
plot_reff_morani <- ggplot(data = d_reff_morani , aes(x = morani_observed, y = reff_mpcr)) +
  geom_smooth(method = "glm", se=FALSE, color="gray80") +
  geom_point(aes(color = outcome_group), pch = 19, size = 2.2, alpha = 0.8) +
  scale_color_manual(values = pcols, guide=guide_legend(title="Outcome group:")) +
  # scale_color_brewer(palette = "Dark2", guide=guide_legend(title="outcome group:")) +
  scale_x_continuous(breaks=seq(0,0.16,by=0.04))+
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  coord_cartesian(xlim=c(-0.01,0.16),ylim=c(1,3.5))+
  facet_grid(country~.) +
  labs(x = "Spatial autocorrelation (Moran's I)",
       y = "",
       tag = "b",
       fill = "Outcome group") +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        # panel.grid.minor.y = element_blank(),
        plot.tag = element_text(face = "bold", size=16),
        strip.text.y = element_blank()
        )

plot_reff_morani

```


```{r plot reff vs prev,warning = FALSE, message = FALSE}

prevplotd <- d_reff_prev %>%
  filter(outcome_lab %in% c("Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm")) %>%
   mutate(
     lab_hjust = case_when(
       country == "Bangladesh" & outcome_lab %in% c("Hookworm","Trichuris sp.","Diarrhea") ~ 0,
       country == "Kenya" & outcome_lab %in% c("Hookworm","Trichuris sp.","Ascaris sp.","Diarrhea") ~ 0,
       TRUE ~ as.numeric(1)
         ),
     lab_xnudge = case_when(
       country == "Bangladesh" & outcome_lab %in% c("Hookworm","Trichuris sp.","Diarrhea") ~ 0.01,
       country == "Kenya" & outcome_lab %in% c("Hookworm","Trichuris sp.","Ascaris sp.","Diarrhea") ~ 0.01,
       TRUE ~ as.numeric(-0.01)
         ),
   )


# pcols <- brewer.pal(n=3,name="Dark2")[3]
plot_reff_prev <- ggplot(data = d_reff_prev %>% filter(outcome_lab %in% c("Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm")), aes(x = Y0bar, y = reff_mpcr)) +
  geom_point(aes(color = outcome_group), pch = 19, size = 2.2, alpha = 0.8) +
  geom_smooth(method = "glm", se=FALSE, color="gray80") +
  scale_color_manual(values = pcols[3], guide=guide_legend(title="Outcome group:")) +
  # scale_fill_brewer(palette = "Dark2", guide=guide_legend(title="outcome group:")) +
  scale_x_continuous(breaks=seq(0,0.4,by=0.1), labels=sprintf("%1.0f",seq(0,0.4,by=0.1)*100))+
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  coord_cartesian(xlim=c(0,0.4),ylim=c(1,3.5))+
  facet_grid(country~.) +
  labs(x = "Outcome prevalence (%)",
       y = "",
       tag = "c",
       fill = "Outcome group") +
  geom_text(aes(label = outcome_lab),
             hjust=prevplotd$lab_hjust, nudge_x = prevplotd$lab_xnudge,
            size = 2.5, show.legend = FALSE, color="gray50") +
  theme_minimal() +
  theme(legend.position = "bottom",
        legend.title = element_text(size = 8),
        legend.text = element_text(size = 8),
        # panel.grid.minor.y = element_blank(),
        plot.tag = element_text(face = "bold", size=16),
        strip.text.y = element_text(size=11, angle = 0, face = "bold"),
        )

plot_reff_prev

```


Create a composite figure

```{r composite icc morani prev figure, fig.width = 8.5, warning = FALSE, message = FALSE}
#https://stackoverflow.com/questions/13649473/add-a-common-legend-for-combined-ggplots
#extract legend
#https://github.com/hadley/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs
g_legend<-function(a_gplot){
  tmp <- ggplot_gtable(ggplot_build(a_gplot))
  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
  legend <- tmp$grobs[[leg]]
  return(legend)
}
complegend<-g_legend(plot_reff_morani)

plot_icc_morani_prev_composite <- grid.arrange(
  arrangeGrob(plot_reff_icc + theme(legend.position="none"),
              plot_reff_morani + theme(legend.position="none"),
              plot_reff_prev + theme(legend.position="none"),
              nrow=1,widths=c(0.7,0.7,1)), # widths=c(0.75,0.96,1)
  complegend, nrow=2,heights=c(10, 1)
)
                                               
                                               # plot_reff_morani,plot_reff_icc,plot_reff_prev,ncol=3,nrow=1,widths=c(0.94,0.94,1))
ggsave(here("output","washb-geopair-reff-icc-morani-prev.pdf"),
       plot_icc_morani_prev_composite,
       device="pdf",
       height=240*0.67,
       width=240,
       units = "mm")
```

# pair matched and sub-district adjusted estimators

Compare relative efficiency of pair matched versus sub-district adjusted estimators of the average treatment effect. 

```{r estimator comparison bangladesh, fig.width = 5, fig.height = 3.2}

# sort by relative efficiency observed for NP (Imai) estimator
orderb <- estcomp_res %>%
  filter(country == "Bangladesh") %>% 
  arrange(reff_mpcr) %>%
  pull(outcome_lab)

estcomp_res2 <- estcomp_res %>%
  mutate(outcome_lab = factor(outcome_lab, levels=orderb))

estcomp_plot_b <-   ggplot(estcomp_res2 %>% filter(country=="Bangladesh")) + 
  geom_hline(yintercept=1, linetype = "dashed") +
  geom_segment( aes(x=outcome_lab, xend=outcome_lab, y=reff_subdreg, yend=reff_mpcr), color="grey60", lwd=0.7) +
  geom_point(aes(x = outcome_lab, y = reff_mpcr), pch=21,
             size = 4, alpha = 1, fill = "black") +
  geom_point(aes(x = outcome_lab, y = reff_subdreg), pch=21,
  size = 4, alpha = 0.9, fill = nytpal2[5]) +
  scale_y_continuous(breaks = seq(1,3.4,by=0.4)) +
  annotate("text",label="Pair matched",y=3.3,x=15, color="black") +
  annotate("text",label="Subdistrict stratified",y=2.1,x=15, color="black") +
  labs(x = NULL, y = "Relative efficiency compared to unmatched analysis",
       tag = "a") +
  coord_flip(ylim = c(1,3.5),xlim=c(1,15)) +
  guides(fill = "none") +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    # panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold", size=16),
    axis.text = element_text(size=12),
    axis.title = element_text(size=12)
        )

estcomp_plot_b
```


```{r estimator comparison kenya, fig.width = 5, fig.height = 3.2}

# sort by relative efficiency observed for pair matched estimator
orderk <- estcomp_res %>%
  filter(country == "Kenya") %>% 
  arrange(reff_mpcr) %>%
  pull(outcome_lab)

estcomp_res2 <- estcomp_res %>%
  mutate(outcome_lab = factor(outcome_lab, levels=orderk))

estcomp_plot_k <-   ggplot(estcomp_res2 %>% filter(country=="Kenya")) + 
  geom_hline(yintercept=1, linetype = "dashed") +
  geom_segment( aes(x=outcome_lab, xend=outcome_lab, y=reff_subdreg, yend=reff_mpcr), color="grey60", lwd=0.8) +
  geom_point(aes(x = outcome_lab, y = reff_mpcr), pch=21,
             size = 4, alpha = 1, fill = "black") +
  geom_point(aes(x = outcome_lab, y = reff_subdreg), pch=21,
             size = 4, alpha = 0.9, fill = nytpal2[5]) +
  scale_y_continuous(breaks = seq(1,3.4,by=0.4)) +
  annotate("text",label="Pair matched",y=3.5,x=13, hjust=1, color="black") +
  annotate("text",label="Subdistrict stratified",y=1.9,x=13, hjust=0.5, color="black") +
  labs(x = NULL, y = "Relative efficiency compared to unmatched analysis",
       tag = "b") +
  coord_flip(ylim = c(1,3.5),xlim=c(1,13)) +
  guides(fill = "none") +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    # panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold", size=16),
    axis.text = element_text(size=12),
    axis.title = element_text(size=12)
        )

estcomp_plot_k

```

Create a composite figure

```{r composite estimator comp figure, fig.height = 9}
plot_estcomp_composite <- grid.arrange(estcomp_plot_b,estcomp_plot_k,ncol=1,nrow=2,heights=c(1,0.85)) # shrink kenya b/c 2 fewer outcomes to keep y-axis in proportion
ggsave(here("output","washb-geopair-estimator-comparison.pdf"),
       plot_estcomp_composite,
       device="pdf",
       height=200,
       width=160, 
       units = "mm")
```

# Session Info
```{r session info}
sessionInfo()
```
