==> R/07-washb-geopair-descriptive-summaries.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "Descriptive summaries"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: pygments
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Create some descriptive tables and figures to summarize key outcomes and study characteristics.

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

# detach MASS and experiment packages
# detach raster and gdistance packages 
# of conflicts with dplyr::select() and not
# needed
detach("package:experiment", unload = TRUE)
detach("package:spaMM", unload = TRUE)
detach("package:MASS", unload = TRUE)
detach("package:gdistance", unload = TRUE)
detach("package:raster", unload = TRUE)


#----------------------------
# source the base functions
#----------------------------
source(here("R","washb-geopair-Functions.R"))
```


# Load data and previous estimates

```{r load indiv level data}
#-------------------------------
# load individual level data
# 
# we need this to calculate age
# summary statistics for each
# outcome (median, range)
#-------------------------------

#-------------------------------
# Bangladesh
# limit to control and nutrition arms
#-------------------------------
# anthropometry data
danth <- read_rds(here("data","bangl_analysis_anthro.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1)) 
  
# child development data
dchdev <- read_rds(here("data","bangl_analysis_chdev.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))

# diarrhea data
ddiar <- read_rds(here("data","bangl_analysis_diar.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))

# parasite data
dpara <- read_rds(here("data","bangl_analysis_parasite.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr=="Control",0,1))



#-------------------------------
# Kenya
# limit to control and nutrition arms
# limit to complete blocks by these two treatments
#-------------------------------
# anthropometry data
dkanth <- read_rds(here("data","kenya_analysis_anthro.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control","Passive Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

# child development data
dkchdev <- read_rds(here("data","kenya_analysis_chdev.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  filter(!is.na(z_easq_com)) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

# diarrhea data
dkdiar <- read_rds(here("data","kenya_analysis_diar.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

# parasite data
dkpara <- read_rds(here("data","kenya_analysis_parasite.rds")) %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  mutate(tr01 = ifelse(tr %in% c("Control"),0,1)) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85))

```


```{r load ate and reff estimates}
#-------------------------------
# load previous estimates of 
# pair-level means and sample
# sizes
#
# matched pair-level data
#-------------------------------
dcl <- read_rds(file = here("output","washb_pair_level_summary_estimates.rds"))

#-------------------------------
# load previous estimates of
# average treatment effects
# and relative efficiency
# 
# outcome-level data
#-------------------------------
ate_res <- read_rds(file = here("output","washb_ATE_releff_estimates.rds"))

#-------------------------------
# load empirical estimates of
# Moran's I for outcomes
#-------------------------------
morani_res <- read_rds(file = here("output","washb-geopair-morans-I-estimates.rds")) %>%
  select(country,outcome_lab,morani_observed)

#-------------------------------
# load empirical estimates of
# the ICC for outcomes
#-------------------------------
icc_res <- read_rds(file = here("output","washb-geopair-icc-estimates.rds")) %>%
  rename(outcome_lab = outcome)
```

# Outcome summary table

```{r outcome summary table data prep}

#-------------------------------
# summarize median age and
# range by outcome
#-------------------------------
get_age <- function(x) {
  agey <- pull(x,agey)
  agemed <- median(agey,na.rm=TRUE)
  agemin <- min(agey, na.rm=TRUE)
  agemax <- max(agey, na.rm=TRUE)
  res <- paste0(sprintf("%1.1f",agemed), " (",sprintf("%1.1f",agemin),", ", sprintf("%1.1f",agemax),")" )
  return(res)
}
bdlist <- list(danth,dchdev,ddiar,dpara)
kdlist <- list(dkanth,dkchdev,dkdiar) # parasite dataset not incl. no age yet
bages <- map_chr(bdlist, get_age)
kages <- map_chr(kdlist, get_age)

# for kenya parasite measuress, temporarily use the 
# weighted avg median reported in Pickering et al. 2019 PLOS Med
# since age is not currently included in the public dataset
kpara_med <- (4928*2 + 4149*5) / (4928+4149)
kages <- c(kages, paste0(sprintf("%1.1f",kpara_med)," (2.0, 15.0)"))

#-------------------------------
# summarize the sample size 
# for each outcome, by group
#-------------------------------
sum_ns <- dcl %>%
  group_by(country, outcome_lab) %>%
  summarize(n0=sum(n0),
            n1=sum(n1),
            .groups="drop"
            )

#-------------------------------
# summarize the SD
# for each outcome, by group
# this requires indiv data
# there is a separate file for
# Bangladesh and Kenya for each
# group of outcomes
#-------------------------------
# anthropometry
sd_anth_b <- danth %>%
  select(tr01,waz,laz,whz,hcz) %>%
  pivot_longer(cols=c(waz,laz,whz,hcz), names_to = "outcome", values_to= "value") %>%
  mutate(country="Bangladesh")
sd_anth_k <- dkanth %>%
  select(tr01,waz,laz,whz,hcz) %>%
  pivot_longer(cols=c(waz,laz,whz,hcz), names_to = "outcome", values_to= "value") %>%
  mutate(country="Kenya")
sd_anth <- sd_anth_b %>%
  bind_rows(sd_anth_k) %>%
  group_by(country,outcome,tr01) %>%
  summarize(sdy = sd(value, na.rm = TRUE), .groups="drop")

# child development
sd_chdev_b <- dchdev %>% 
  select(tr01, starts_with("z_")) %>%
  pivot_longer(cols= starts_with("z_"), names_to = "outcome", values_to= "value") %>%
  mutate(country="Bangladesh")
sd_chdev_k <- dkchdev %>% 
  select(tr01, starts_with("z_")) %>%
  pivot_longer(cols= starts_with("z_"), names_to = "outcome", values_to= "value") %>%
  mutate(country="Kenya")
sd_chdev <- sd_chdev_b %>%
  bind_rows(sd_chdev_k) %>%
  group_by(country, outcome,tr01) %>%
  summarize(sdy = sd(value, na.rm = TRUE), .groups="drop") %>%
  filter(outcome != "z_easq_total")

# diarrhea
sd_diar_b <- ddiar %>%
  select(tr01, diar7d) %>%
  mutate(country = "Bangladesh")
sd_diar_k <- dkdiar %>%
  select(tr01, diar7d) %>%
  mutate(country = "Kenya")
sd_diar <- sd_diar_b %>%
  bind_rows(sd_diar_k) %>%
  mutate(outcome = "diar7d") %>%
  rename(value=diar7d) %>%
  group_by(country, outcome,tr01) %>%
  summarize(sdy = sd(value, na.rm = TRUE), .groups="drop")

# parasite infection
sd_para_b <- dpara %>%
  select(tr01, al, tt, hw, giar) %>%
  pivot_longer(cols= c(al, tt, hw, giar), names_to = "outcome", values_to= "value") %>%
  mutate(country = "Bangladesh")
sd_para_k <- dkpara %>%
  select(tr01, al, tt, hw, giar) %>%
  pivot_longer(cols= c(al, tt, hw, giar), names_to = "outcome", values_to= "value") %>%
  mutate(country = "Kenya")
sd_para <- sd_para_b %>%
  bind_rows(sd_para_k) %>%
  group_by(country, outcome,tr01) %>%
  summarize(sdy = sd(value, na.rm = TRUE), .groups="drop")

# stack all of the SD estimates together
# pivot wider for treatment groups
# format outcome labels so can merge to 
# the ATE estimates and all other data
sd_ests <- bind_rows(sd_anth,sd_chdev,sd_diar,sd_para) %>%
  pivot_wider(id_cols=c(country,outcome), names_from = tr01, names_prefix = "sd", values_from="sdy") %>%
  mutate(outcome_lab = case_when(
    outcome == "laz" ~ "length-for-age z",
    outcome == "waz" ~ "weight-for-age z",
    outcome == "whz" ~ "weight-for-height z",
    outcome == "hcz" ~ "head circumference z",
    outcome == "z_easq_com" ~ "EASQ communication z",
    outcome == "z_easq_motor" ~ "EASQ gross motor z",
    outcome == "z_easq_pers" ~ "EASQ personal-social z",
    outcome == "z_cdi_comp" ~ "CDI comprehension z",
    outcome == "z_cdi_expr" ~ "CDI expression z",
    outcome == "diar7d" ~ "Diarrhea",
    outcome == "giar" ~ "Giardia sp.",
    outcome == "al" ~ "Ascaris sp.",
    outcome == "tt" ~ "Trichuris sp.",
    outcome == "hw" ~ "Hookworm"
  ),
  outcome_lab = factor(outcome_lab, levels=levels(dcl$outcome_lab))
         ) %>%
  select(-outcome)

#-------------------------------
# join the various summaries together 
# into a single data.frame
#-------------------------------
outtab <- ate_res %>%
  mutate(ate_print = paste0(sprintf("%1.2f",est)," (", sprintf("%1.2f",est_min95), ", ", sprintf("%1.2f",est_max95),")" )) %>%
  left_join(sum_ns,by=c("country","outcome_lab")) %>%
  left_join(sd_ests, by = c("country","outcome_lab")) %>%
  left_join(icc_res,by=c("country","outcome_lab")) %>%
  left_join(morani_res,by=c("country","outcome_lab")) %>%
  
  # add in ages (super clunky!)
  mutate(age_print = case_when(
    country=="Bangladesh" & outcome_lab %in% c("length-for-age z","weight-for-age z", "weight-for-height z", "head circumference z") ~ bages[1],
    country=="Bangladesh" & outcome_lab %in% c("EASQ communication z","EASQ gross motor z", "EASQ personal-social z","CDI comprehension z","CDI expression z") ~ bages[2],
    country=="Bangladesh" & outcome_lab %in% c("Diarrhea") ~ bages[3],
    country=="Bangladesh" & outcome_lab %in% c("Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm") ~ bages[4],
    
    country=="Kenya" & outcome_lab %in% c("length-for-age z","weight-for-age z", "weight-for-height z", "head circumference z") ~ kages[1],
    country=="Kenya" & outcome_lab %in% c("EASQ communication z","EASQ gross motor z", "EASQ personal-social z") ~ kages[2],
    country=="Kenya" & outcome_lab %in% c("Diarrhea") ~ kages[3],
    country=="Kenya" & outcome_lab %in% c("Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm") ~ kages[4]
  )
         ) %>%
  select(country,outcome_lab, age_print,
         n0,Y0bar,sd0, n1,Y1bar,sd1,
         ate_print,
         icc_print,
         morani=morani_observed
         )


```



```{r outcome summary table bangladesh, out.width=12}
#-------------------------------
# outcome summary table
# Bangladesh
#-------------------------------
options(knitr.kable.NA = "‡")
knitr::kable(outtab %>% filter(country=="Bangladesh") %>% select(-country),
             digits = c(0,0,0,2,2,0,2,2,0,0,2),
             align = c("lcrccrccccc"),
             col.names = c("Outcome","Median age y (range)", "n","mean","sd","n","mean","sd","ATE (95% CI)*", "ICC (95% CI)†", "Moran's I"),
             caption = "Summary of outcomes in the WASH Benefits Bangladesh trial",
             format.args = list(big.mark = ",")
             ) %>%
  kable_styling(bootstrap_options = "striped") %>%
  pack_rows("Child growth",1,4) %>%
  pack_rows("Child development",5,9) %>%
  pack_rows("Infectious disease",10,14) %>%
  add_header_above(c(" " = 2, "Control" = 3, "Nutrition" = 3, " "=3)) %>%
  footnote(symbol = c("ATE: Average treatment effect of difference in means (Nutrition minus Control), with 95% confidence interval reported for the non-parametric, pair-matched estimator. For infectious disease outcomes, this is the difference in the proportion positive.",
                      "ICC: Intra-cluster correlation coefficient among outcomes measured in the control group, with 95% confidence interval estimated through bootstrap resampling geographic pairs with replacement (1,000 iterations)."))

```


```{r outcome summary table kenya, out.width=12}

knitr::kable(outtab %>% filter(country=="Kenya") %>% select(-country),
             digits = c(0,0,0,2,2,0,2,2,0,0,2),
             align = c("lcrccrccccc"),
             col.names = c("Outcome","Median age y (range)", "n","mean","sd","n","mean","sd","ATE (95% CI)*", "ICC (95% CI)†", "Moran's I"),
             caption = "Summary of outcomes in the WASH Benefits Kenya trial",
             format.args = list(big.mark = ",")
             ) %>%
  kable_styling(bootstrap_options = "striped") %>%
  pack_rows("Child growth",1,4) %>%
  pack_rows("Child development",5,7) %>%
  pack_rows("Infectious disease",8,12) %>%
  add_header_above(c(" " = 2, "Control" = 3, "Nutrition" = 3, " "=3)) %>%
  footnote(symbol = c("ATE: Average treatment effect of difference in means (Nutrition minus Control), with 95% confidence interval reported for the non-parametric, pair-matched estimator. For infectious disease outcomes, this is the difference in the proportion positive.",
                      "ICC: Intra-cluster correlation coefficient among outcomes measured in the control group, with 95% confidence interval estimated through bootstrap resampling geographic pairs with replacement (1,000 iterations).","The mixed effects model did not converge for Trichuris sp. infection in Kenya, so no estimate of the ICC was possible.", "MacArthur-Bates Communicative Development Inventory (CDI) measures were not collected in the Kenya trial."))

```


# Correlation and relative efficiency table

```{r summary relative efficacy table}
options(knitr.kable.NA = "‡")

sumtab <- ate_res %>%
  mutate(print_corr = paste0(sprintf("%1.2f",corr_w)," (",sprintf("%1.2f",cor_min95),", ",sprintf("%1.2f",cor_max95),")"),
         print_reff = paste0(sprintf("%1.1f",eff_w)," (",sprintf("%1.1f",releff_min95),", ",sprintf("%1.1f",releff_max95),")"),
         )

sumtab <- sumtab %>%
  filter(country=="Bangladesh") %>%
  select(outcome_lab, bcorr = print_corr, beff = print_reff) %>%
  left_join(sumtab %>% filter(country=="Kenya"), by = "outcome_lab") %>%
  select(outcome_lab, bcorr, kcorr=print_corr, beff, keff=print_reff)


knitr::kable(sumtab,
             digits = c(0,2,2,1,1),
             align = c("lcccc"),
             col.names = c("Outcome","Bangladesh","Kenya","Bangladesh","Kenya"),
             caption = "Summary of pair-wise outcome correlation and relative efficacy of geographic pair-matching") %>%
  kable_styling(bootstrap_options = "striped") %>%
  pack_rows("Child growth",1,4) %>%
  pack_rows("Child development",5,9) %>%
  pack_rows("Infectious disease",10,14) %>%
  add_header_above(c(" " = 1, "Pair-wise correlation (95% CI)*" = 2, "Relative Efficiency (95% CI)†" = 2)) %>%
  footnote(symbol = c("Weighted correlation between geographically paired, cluster-level means. 95% confidence interval estimated using a non-parametric bootstrap of pairs.",
                      "Relative efficiency of geographic pair-matching compared to an unmatched design, defined as 1/(1-r), where r is the weighted correlation between paired outcomes. 95% confidence interval estimated using a non-parametric bootstrap of pairs.","MacArthur-Bates Communicative Development Inventory (CDI) measures were not collected in the Kenya trial."))

```

# Bangladesh cluster sizes

```{r cluster size data prep}
#---------------------------------
# create an outcome group var
#
# create a shorter outcome group
# label
#
# calculate difference in pair
# level sample sizes
#---------------------------------
dcl2 <- dcl %>%
  mutate(outcome_group = case_when(
    outcome_lab %in% c("length-for-age z","weight-for-age z","weight-for-height z", "head circumference z") ~ "Child growth",
    outcome_lab %in% c("Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm") ~ "Infectious disease",
    TRUE ~ "Child development"
  ),
  outcome_group = factor(outcome_group),
  outcome_lab_short = case_when(
    outcome_lab == "length-for-age z" ~ "LAZ",
    outcome_lab == "weight-for-age z" ~ "WAZ",
    outcome_lab == "weight-for-height z" ~ "WHZ",
    outcome_lab == "head circumference z" ~ "HCZ",
    outcome_lab == "EASQ communication z" ~ "EASQ com.",
    outcome_lab == "EASQ gross motor z" ~ "EASQ motor",
    outcome_lab == "EASQ personal-social z" ~ "EASQ social",
    outcome_lab == "CDI comprehension z" ~ "CDI comp.",
    outcome_lab == "CDI expression z" ~ "CDI expr.",
    TRUE ~ as.character(outcome_lab)
  ),
  outcome_lab_short = factor(outcome_lab_short,
                             levels=c("LAZ","WAZ","WHZ","HCZ","EASQ com.","EASQ motor","EASQ social","CDI comp.","CDI expr.","Diarrhea","Giardia sp.","Ascaris sp.","Trichuris sp.","Hookworm")),
  diffn = n1-n0
         )


#---------------------------------
# pivot the pair-level data to 
# long by treatment for plotting
#---------------------------------
dcl_long <- dcl2 %>%
  select(country, outcome_group, outcome_lab_short, n0, n1) %>%
  pivot_longer(cols=c("n0","n1"),names_prefix = "n",names_to = "tx",values_to = "n") %>%
  mutate(txlab = ifelse(tx==1,"Nutr.","Control"),
         txlab = factor(txlab)
  )
```

```{r cluster size bangladesh growth}
#-------------------------------
# Bangladesh
# cluster sizes
# growth
#-------------------------------
# pcols <- cbpal[2:3]
pcols <- cbpal[c(5,4)]
# pcols <- c("gray70",cbpal[7])
# vircols <- viridis(n=9)
# pcols <- c("gray70",vircols[2])

plot_cluster_sizes_growth_b <- ggplot(data=dcl_long %>% filter(country=="Bangladesh" & outcome_group=="Child growth"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="a") +
  scale_y_continuous(breaks=seq(8,16,by=2)) +
  coord_cartesian(ylim=c(8,16)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster size bangladesh chdev}
#-------------------------------
# Bangladesh
# cluster sizes
# child development
#-------------------------------
plot_cluster_sizes_chdev_b <- ggplot(data=dcl_long %>% filter(country=="Bangladesh" & outcome_group=="Child development"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="b") +
  scale_y_continuous(breaks=seq(6,16,by=2)) +
  coord_cartesian(ylim=c(6,16)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster size bangladesh ID}
#-------------------------------
# Bangladesh
# cluster sizes
# infectious disease
#-------------------------------
plot_cluster_sizes_infd_b <- ggplot(data=dcl_long %>% filter(country=="Bangladesh" & outcome_group=="Infectious disease"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="c") +
  scale_y_continuous(breaks=seq(0,70,by=10)) +
  coord_cartesian(ylim=c(0,70)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```


```{r cluster diff bangladesh growth}
#-------------------------------
# Bangladesh 
# difference in pair-level
# cluster sizes, growth
#-------------------------------
plot_cluster_diff_growth_b <- ggplot(data=dcl2 %>% filter(country=="Bangladesh" & outcome_group=="Child growth"), aes(x = diffn,)) +
  geom_histogram(binwidth=1,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 4) +
  scale_y_continuous(breaks=seq(0,16,by=4))+
  scale_x_continuous(breaks=seq(-6,6,by=2))+
  coord_flip(xlim=c(-6,6),ylim=c(0,18))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="d")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster diff bangladesh chdev}
#-------------------------------
# Bangladesh 
# difference in pair-level
# cluster sizes, child development
#-------------------------------
plot_cluster_diff_chdev_b <- ggplot(data=dcl2 %>% filter(country=="Bangladesh" & outcome_group=="Child development"), aes(x = diffn,)) +
  geom_histogram(binwidth=1,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 5) +
  scale_y_continuous(breaks=seq(0,24,by=6))+
  scale_x_continuous(breaks=seq(-10,6,by=2))+
  coord_flip(xlim=c(-10,6),ylim=c(0,24))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="e")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster diff bangladesh ID}
#-------------------------------
# Bangladesh 
# difference in pair-level
# cluster sizes, infectious disease
#-------------------------------
plot_cluster_diff_infd_b <- ggplot(data=dcl2 %>% filter(country=="Bangladesh" & outcome_group=="Infectious disease"), aes(x = diffn,)) +
  geom_histogram(binwidth=3,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 5) +
  scale_y_continuous(breaks=seq(0,24,by=8))+
  scale_x_continuous(breaks=seq(-30,30,by=6))+
  coord_flip(xlim=c(-30,30),ylim=c(0,24))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="f")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster composite plot bangladesh, fig.width=12, fig.height=12}
#-------------------------------
# Bangladesh
# cluster sizes
# composite plot
#-------------------------------
plot_cluster_b_composite <- grid.arrange(plot_cluster_sizes_growth_b,
                                         plot_cluster_diff_growth_b,
                                         plot_cluster_sizes_chdev_b,
                                         plot_cluster_diff_chdev_b,
                                         plot_cluster_sizes_infd_b,
                                         plot_cluster_diff_infd_b,
                                         ncol=2,nrow=3
                                         )
ggsave(here("output","washb-geopair-cluster-sizes-bangl.png"),plot_cluster_b_composite,device="png",height=10,width=10)

```



# Kenya cluster sizes

```{r cluster size Kenya growth}
#-------------------------------
# Kenya
# cluster sizes
# growth
#-------------------------------
plot_cluster_sizes_growth_k <- ggplot(data=dcl_long %>% filter(country=="Kenya" & outcome_group=="Child growth"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="a") +
  scale_y_continuous(breaks=seq(0,30,by=5)) +
  coord_cartesian(ylim=c(3,33)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```



```{r cluster size kenya chdev}
#-------------------------------
# Kenya
# cluster sizes
# child development
#-------------------------------
plot_cluster_sizes_chdev_k <- ggplot(data=dcl_long %>% filter(country=="Kenya" & outcome_group=="Child development"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="b") +
  scale_y_continuous(breaks=seq(0,30,by=5)) +
  coord_cartesian(ylim=c(3,30)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster size kenya ID}
#-------------------------------
# Kenya
# cluster sizes
# infectious disease
#-------------------------------
plot_cluster_sizes_infd_k <- ggplot(data=dcl_long %>% filter(country=="Kenya" & outcome_group=="Infectious disease"),
                               aes(x= txlab, y = n, color = txlab, fill=txlab)) +
  geom_violin(color="gray30", alpha=0.6, draw_quantiles = c(0.50)) +
  # facet over outcome
  facet_wrap(~outcome_lab_short,ncol = 5) +
  # aesthetics
  scale_fill_manual(values=pcols) +
  scale_color_manual(values=pcols) +
  labs(x="",y="Number of children per matched pair",tag="c") +
  scale_y_continuous(breaks=seq(0,60,by=10)) +
  coord_cartesian(ylim=c(0,65)) +
  theme_minimal() +
  theme(
    legend.position = "none",
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```


```{r cluster diff kenya growth}
#-------------------------------
# Kenya
# pair-level difference in 
# cluster sizes, growth
#-------------------------------
plot_cluster_diff_growth_k <- ggplot(data=dcl2 %>% filter(country=="Kenya" & outcome_group=="Child growth"), aes(x = diffn,)) +
  geom_histogram(binwidth=2,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 4) +
  scale_y_continuous(breaks=seq(0,16,by=4))+
  scale_x_continuous(breaks=seq(-12,14,by=4))+
  coord_flip(xlim=c(-12,14),ylim=c(0,16))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="d")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )
```

```{r cluster diff kenya chdev}
#-------------------------------
# Kenya
# pair-level difference in 
# cluster sizes, child development
#-------------------------------
plot_cluster_diff_chdev_k <- ggplot(data=dcl2 %>% filter(country=="Kenya" & outcome_group=="Child development"), aes(x = diffn,)) +
  geom_histogram(binwidth=2,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 5) +
  scale_y_continuous(breaks=seq(0,16,by=4))+
  scale_x_continuous(breaks=seq(-12,14,by=4))+
  coord_flip(xlim=c(-12,14),ylim=c(0,16))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="e")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )
```



```{r cluster diff kenya ID}
#-------------------------------
# Kenya
# pair-level difference in 
# cluster sizes, infectious disease
#-------------------------------
plot_cluster_diff_infd_k <- ggplot(data=dcl2 %>% filter(country=="Kenya" & outcome_group=="Infectious disease"), aes(x = diffn,)) +
  geom_histogram(binwidth=3,fill=cbpal[6],alpha=0.5,color="gray40") +
  geom_vline(xintercept=0,color="gray20", linetype = "dashed") +
  facet_wrap(~outcome_lab_short,ncol = 5) +
  scale_y_continuous(breaks=seq(0,16,by=4))+
  scale_x_continuous(breaks=seq(-18,30,by=6))+
  coord_flip(xlim=c(-21,33),ylim=c(0,16))+
  labs(y="Number of pairs",x = "Diff in sample size (nutrition - control)",tag="f")+
  theme_minimal() +
  theme(
    legend.position = "none",
    panel.grid.minor = element_blank(),
    plot.tag = element_text(face = "bold",size=16),
    strip.text = element_text(size=10)
  )

```

```{r cluster composite plot kenya, fig.width=12, fig.height=12}
#-------------------------------
# Kenya
# cluster size composite plot
#-------------------------------
plot_cluster_k_composite <- grid.arrange(plot_cluster_sizes_growth_k,
                                         plot_cluster_diff_growth_k,
                                         plot_cluster_sizes_chdev_k,
                                         plot_cluster_diff_chdev_k,
                                         plot_cluster_sizes_infd_k,
                                         plot_cluster_diff_infd_k,
                                         ncol=2,nrow=3
                                         )
ggsave(here("output","washb-geopair-cluster-sizes-kenya.png"),plot_cluster_k_composite,device="png",height=10,width=10)

```


# Session Info
```{r session info}
sessionInfo()
```
