==> R/09-washb-geopair-relative-efficiency-across-scales.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials."
subtitle: "Relative efficiency  across geographic scales"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: haddock
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

In this analysis, we took advantage of the large scale of the Bangladesh and Kenya trials to study the relative efficiency of pair matching at different trial sizes, which correspond with different geographic scales. We studied releative efficiency of different trials sizes from 10 matched pairs up to the full trial size in each country (90 in Bangladesh, 72 in Kenya) down-sample the matched pairs

# Preamble

```{r preamble}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

# detach gdistance, raster, experiment, MASS  packages 
# because of conflicts with dplyr::select() and not
# needed
detach("package:gdistance", unload = TRUE)
detach("package:raster", unload = TRUE)
detach("package:experiment", unload = TRUE)
detach("package:spaMM", unload = TRUE)
detach("package:MASS", unload = TRUE)

#----------------------------
# Box data directory
#----------------------------
Box_data_directory <- "~/Library/CloudStorage/Box-Box/washb-geopair/data/"


```




# Data prep

```{r load data}
#-------------------------------
# load the pair-level summary
# estimates, created by
# washb-geopair-relative-efficiency.R
#-------------------------------
dcl <- read_rds(here("output","washb_pair_level_summary_estimates.rds"))

#-------------------------------
# load the distance matrix
# between control cluster block
# centroids
# created by
# washb-geopair-gps-data-processing.R
#-------------------------------
distmat_b <- read_rds(here(Box_data_directory,"final/bangl_analysis_block_dists.rds"))

kblocks <- unique(dcl$block[dcl$country=="Kenya" & dcl$outcome_lab=="Ascaris sp."])

distmat_k <- read_rds(here(Box_data_directory,"final/kenya_analysis_block_dists.rds")) %>%
  # filter out blocks that were not complete for control,N, N+WSH
  # ATTENTION:
  # this should probably be done earlier, in the GPS data processing workflow.
  filter(block %in% kblocks)  %>%
  group_by(block) %>%
  pivot_longer(cols=starts_with("block_"),
               names_prefix="block_",
               names_to = "blockd", 
               values_to = "dist"
               ) %>%
  filter(blockd %in% kblocks) %>%
  pivot_wider(id_cols=block,names_from="blockd",values_from="dist",names_prefix="block_") %>%
  ungroup() %>%
  data.frame()


#-------------------------------
# create outcome groups
# for plotting
#-------------------------------
dcl <- dcl %>%
  mutate(outcome_group = case_when(
    outcome_lab %in% c("length-for-age z","weight-for-age z","weight-for-height z","head circumference z") ~ "Child growth",
    outcome_lab %in% c("EASQ communication z","EASQ gross motor z","EASQ personal-social z","CDI comprehension z","CDI expression z") ~ "Child development",
    TRUE ~ as.character("Infectious disease")
  ),
  outcome_group = factor(outcome_group)
         )

#-------------------------------
# split data by country, since
# analyses will proceed separately
# by country
#-------------------------------
dcl_b <- dcl %>% filter(country=="Bangladesh")

dcl_k <- dcl %>% filter(country=="Kenya") %>% 
  mutate(outcome_lab = factor(outcome_lab)) #%>%

```

# Geographic sampling functions

```{r geographic sampling function}
#-------------------------------
# geosample()
# function to randomly sample 
# a block, and then identify n
# geographically closest blocks 
# to the sampled block
#
# @distframe :  a data frame with pair-wise distances between blocks (first col is block ID)
# @nclosest  : number of closest blocks to identify
# 
# returns a list with two objects:
# $block_select : IDs of blocks selected
# $block_dist   : distance in meters between selected blocks
#-------------------------------
geosample <- function(distframe,nclosest) {
  
  # store block IDs
  block_ids <- distframe$block
  block_dists <- distframe %>% select(-block)
  
  # sample a random column from distframe
  samp_col <- sample(block_ids,size=1)
  samp_col <- paste0("block_",samp_col)
  
  # identify the n closest blocks to the 
  # randomly sampled block
  x_samp <- distframe %>% select(block,all_of(samp_col))
  x_ord <- order(x_samp %>% pull(samp_col))
  selected_blocks <- x_samp[x_ord,][1:(nclosest),] %>% pull(block)
  selected_dists  <- x_samp[x_ord,][1:(nclosest),] %>% pull(samp_col)
  
  list(block_select = selected_blocks, block_dist = selected_dists )
}

#-------------------------------
# geofind()
# similar to geosample() (above)
# but rather than randomly sampling
# a block, simply identify n
# geographically closest blocks 
# to an identified block this is 
# useful in an exhaustive search
# across all blocks
#
# this function is not actually
# used here, but left for posterity
#
# @distframe : a data frame with pair-wise distances between blocks (first col is block ID)
# @blockid   : ID of the block selected
# @nclosest  : number of closest blocks to identify
# 
# returns a list with two objects:
# $block_select : IDs of blocks selected
# $block_dist   : distance in meters between selected blocks
#-------------------------------
geofind <- function(distframe,blockid,nclosest) {
  
  # identify the column specified for the particular block
  samp_col <- paste0("block_",blockid)
  
  # identify the n closest blocks to the 
  # randomly sampled block
  x_samp <- distframe %>% 
    select(block,all_of(samp_col)) %>%
    filter(!block %in% blockid)
  x_ord <- order(x_samp %>% pull(samp_col))
  selected_blocks <- x_samp[x_ord,][1:(nclosest),] %>% pull(block)
  selected_dists  <- x_samp[x_ord,][1:(nclosest),] %>% pull(samp_col)
  
  list(block_select = selected_blocks, block_dist = selected_dists )
}


```

# Bangladesh

## Correlation across scales

For each outcome, estimate the correlation between paired means.

Repeat estimates from 10 up to 90 pairs. 

For geographically proximate samples, compute estimates based on all 90 blocks selected as the initial sampling point, in turn (an exaustive sample).


```{r bangladesh pairwise correlation by number of pairs}
#-------------------------------
# Bangladesh
#
# Correlation estimates at
# different trial sizes, 
# resampling index clusters
# with replacement and selecting
# all n-1 pairs in closest 
# proximity
#-------------------------------
start_time <- Sys.time()
cor_geo_b <- foreach(outi = levels(dcl_b$outcome_lab), .combine = rbind) %do% {
  foreach(npairs = seq(10,90,by=2), .combine = rbind) %:% 
  foreach(booti = 1:1000, .combine = rbind) %dopar% {
    
    # set seed within %dopar% for perfect reproducibility
    set.seed(booti)
    
    # randomly sample a block and select the n closest pairs
    # identify the geographically n proximate npairs to
    sampi <- geosample(distframe = distmat_b, nclosest = npairs)
    
    # then draw a bootstrap resample from the set of n closest pairs
    bsampi <- sample(1:npairs,size=npairs,replace=TRUE)
    bsamp  <- sampi$block_select[bsampi]
    bdist  <- sampi$block_dist[bsampi]
    
    di <- data.frame(block=as.character(bsamp), outcome_lab = outi) %>%
      left_join(dcl_b, by=c("outcome_lab","block"))
    
    # estimate outcome correlation, weighted by pair-wise N
    wY1bar <- di$Y1bar*di$w
    wY0bar <- di$Y0bar*di$w
    corri <- 2*cov(wY1bar,wY0bar) / ( var(wY1bar) + var(wY0bar) )
    
    # save the minimum and maximum distances between clusters
    mindisti <- min(bdist[as.numeric(bdist)>0])
    maxdisti <- max(bdist)
    
    # return results
    res <- data.frame(outcome_lab = outi, npairs , mindist = mindisti, maxdist = maxdisti, corr = corri)
    res
  }
}


cor_ests_b <- cor_geo_b %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl_b$outcome_lab)))

#----------------------------
# check time required to
# run bootstrap reps
#----------------------------
end_time <- Sys.time()
end_time - start_time
```


```{r bangladesh estimate mean correlation}
#-------------------------------
# estimate the mean and 95% CI
# for correlation across
# the different resampling
# replicates
#-------------------------------
sum_ests_b <- cor_ests_b %>% 
  group_by(outcome_lab,npairs) %>% 
  filter(!is.nan(corr)) %>%
  summarize(cor_est = mean(corr), 
            cor_min95 = quantile(corr,probs = 0.025, na.rm=TRUE),
            cor_max95 = quantile(corr,probs = 0.975, na.rm=TRUE),
            .groups = "keep") %>%
  mutate(x = npairs/2)

#-------------------------------
# estimate the average 
# maximum distance between
# clusters across
# the different resampling
# replicates
# this is the same for every
# outcome so can simplify
#-------------------------------
maxdist_ests_b <- cor_ests_b %>%
  # restrict to just one outcome (arbitrary)
  filter(outcome_lab == "Diarrhea") %>%
  group_by(npairs) %>% 
  summarize(maxdist_mu = mean(maxdist),
            maxdist_lb = quantile(maxdist,prob=0.025),
            maxdist_ub = quantile(maxdist,prob=0.975),
            .groups = "drop"
  ) %>%
  # convert distances to numeric and from m to km
  mutate(maxdist_mu = as.numeric(maxdist_mu)/1000,
         maxdist_lb = as.numeric(maxdist_lb)/1000,
         maxdist_ub = as.numeric(maxdist_ub)/1000
         )
```


## Estimate 95% CIs for relative efficiency

Use a bootstrap resampling approach, resampling pairs with replacement, to estimate the 95% confidence interval for relative efficiency of geographic pair matching with 90 matched pairs. This will provide a sense for whether lower efficiency at smaller sample sizes is significantly lower, and not just a function of sampling variability.

```{r bangladesh bootstrap CIs for relative efficiency}
#-------------------------------
# pivot wider to make the bootstrap
# easy
#-------------------------------
dboot <- dcl_b %>%
  group_by(outcome_lab)

# estimate outcome correlation, weighted by pair-wise N
corx <- function(x,y,w) {
    xw <- x*w
    yw <- y*w
    wcorr <- 2*cov(xw,yw) / ( var(xw) + var(yw) )
    return(wcorr)
}

#-------------------------------
# boostrap resample within
# each outcome group
# 1000 iterations
#-------------------------------
cor_boot_b <- foreach(iteri = 1:1000, .combine = rbind) %dopar% {
  set.seed(iteri)
  di <- dboot %>% 
    slice_sample(prop = 1, replace = TRUE)
  corest <- di %>% group_map( ~ corx(x=.x$Y1bar, y=.x$Y0bar,w = .x$w))
  res <- data.frame(iter = iteri, outcome_lab = levels(di$outcome_lab), corr = unlist(corest))
  res
}

cor_boot_b2 <- cor_boot_b %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl_b$outcome_lab)))

#-------------------------------
# calculate percentile 95% CIs 
# from the bootstrap distribution
#-------------------------------
cor_boot_ci_b <- cor_boot_b2 %>%
  group_by(outcome_lab) %>%
  mutate(cor_min95 = quantile(corr, probs = 0.025, na.rm=TRUE),
         cor_max95 = quantile(corr, probs = 0.975, na.rm=TRUE)) %>%
  slice(1) %>%
  select(outcome_lab, cor_min95, cor_max95)

#-------------------------------
# transform correlation 95% CIs 
# into relative efficiency
#-------------------------------
releff_ci_b <- cor_boot_ci_b %>%
  mutate(releff_min95 = 1/(1-cor_min95),
         releff_max95 = 1/(1-cor_max95))

```

## Supplementary Figure 7 corr by size

Plot correlation and bootstrapped 95% CI by down-sampled trial size

```{r plot correlation bangladesh}

# calculate relative efficacy from correlation
dreleff_b <- sum_ests_b %>%
  mutate(releff = 1/(1-cor_est),
         releff_lb = 1/(1-cor_min95),
         releff_ub = 1/(1-cor_max95)) %>%
  # merge on the outcome groups 
  left_join( dcl_b %>% group_by(outcome_lab) %>% slice(1) %>% select(outcome_lab,outcome_group), by="outcome_lab") %>%
  # re-arrange order of the outcome groups for plot aesthetics
  mutate(outcome_lab_rev = factor(outcome_lab, levels = rev(levels(dcl$outcome_lab))) )

# plot correlation with bootstrap 95% CI
pcols <- c(cbpal[2],cbpal[4],"gray50")
plot_corr <- ggplot(data=dreleff_b, aes(x = npairs, color = outcome_group, fill = outcome_group)) +
  # facet over outcome
  facet_wrap(~outcome_lab_rev,ncol = 5) +
  
  # plot results
  # geom_hline(data=releff_ci_b_plot2,aes(yintercept=releff_min95),color="gray40",lty="dashed") +
  geom_hline(yintercept = 0, color="gray60", linetype = "dashed") +
  geom_ribbon(aes(ymin = cor_min95, ymax = cor_max95), alpha=0.2, color = NA) +
  geom_line(aes(y = cor_est)) +
  # refine aesthetics
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  # scale_color_brewer(palette = "Dark2") +
  # scale_fill_brewer(palette = "Dark2") +
  scale_y_continuous(breaks=seq(-0.6,1,by=0.2), labels=paste0(sprintf("%1.1f",seq(-0.6,1,by=0.2))) )+
  scale_x_continuous(breaks=seq(10,90,by=20)) +
  coord_cartesian(xlim=c(10,90),ylim=c(-0.6,1)) +
  labs(x="Down sampled trial size: number of geographically proximate matched pairs",y="Pair-level outcome correlation (r)") +
  guides(colour = guide_legend(title = "N pairs",nrow = 1)) +
  theme_minimal() +
  theme(
    legend.position = "none"  
    )
plot_corr

ggsave(here("output","washb-geopair-corr-geographic-scale-smallreps-bangladesh.png"), plot_corr, device = "png", width = 8, height = 8)
```

## Relative efficiency across geographic scales

```{r plot relative efficiency bangladesh compressed}
#-------------------------------
# calculate relative efficacy 
# from correlation
#-------------------------------
dreleff_b <- sum_ests_b %>%
  mutate(releff = 1/(1-cor_est),
         releff_lb = 1/(1-cor_min95),
         releff_ub = 1/(1-cor_max95)) %>%
  # merge on the outcome groups 
  left_join( dcl_b %>% group_by(outcome_lab) %>% slice(1) %>% select(outcome_lab,outcome_group), by="outcome_lab") %>%
  # re-arrange order of the outcome groups for plot aesthetics
  mutate(outcome_lab_rev = factor(outcome_lab, levels = rev(levels(dcl$outcome_lab))) )

#-------------------------------
# add correlation at 90 pairs
# to color each outcome in the
# figure
#-------------------------------
cor90ests_b <- dreleff_b %>%
  ungroup() %>%
  filter(npairs == 90) %>%
  select(outcome_group,outcome_lab, cor90 = cor_est, releff90 = releff) %>%
  # small aesthetic tweak in ordering
  # laz is higher than waz for the entire series
  # until the very last measurement. bump it up a tiny
  # bit so laz appears higher in the ordering
  mutate(releff90=ifelse(outcome_lab=="length-for-age z",releff90+0.2,releff90)) %>%
  arrange(outcome_group, desc(releff90))

dreleff_b <- dreleff_b %>%
  left_join(cor90ests_b, by = c("outcome_group","outcome_lab"))

#-------------------------------
# store outcome labels to add to
# plot
#-------------------------------
out_labs_print <- cor90ests_b %>%
  group_by(outcome_group) %>%
  arrange(outcome_group, desc(releff90)) %>%
  mutate(outcome_order = seq_along(outcome_lab)) %>%
  mutate(y_text = 3.65 - outcome_order*0.15) %>% 
  ungroup() %>%
  mutate(releff_cat = case_when(
    releff90 <1.4 ~ "<1.4",
    releff90 >= 1.4 & releff90 < 1.8 ~ "1.4 to 1.8",
    releff90 >= 1.8 & releff90 < 2.2 ~ "1.8 to 2.2",
    releff90 >= 2.2 & releff90 < 2.6 ~ "2.2 to 2.6",
    releff90 >= 2.6 & releff90 < 3.0 ~ "2.6 to 3.0",
    releff90 >3.0 ~ ">3.0",
    TRUE ~ as.character(NA)
  ),
  releff_cat = factor(releff_cat, levels=c("<1.4","1.4 to 1.8","1.8 to 2.2","2.2 to 2.6","2.6 to 3.0",">3.0"))
         ) %>%
  select(outcome_group,outcome_lab,releff_cat, outcome_order, y_text)
  
dreleff_b <- dreleff_b %>%
  left_join(out_labs_print, by = c("outcome_group","outcome_lab"))

#-------------------------------
# make plot
#-------------------------------
# pcols <- rev(nytpal[2:7])
# pcols <- c("gray40",nytpal[c(7,6,4,3,2)])
pcols <- viridis(n=6,begin=0.1,end=0.8,option="D",direction=1)

plot_reff_b <- ggplot(data=dreleff_b, aes(x = npairs, y = releff, color = releff_cat, fill = releff_cat, group = outcome_lab)) +
  # facet over outcome
  facet_grid(~outcome_group) +
  # plot results
  # geom_ribbon(aes(ymin = releff_lb, ymax = releff_ub), alpha=0.2, color = NA, show.legend = FALSE) +
  geom_line() +
  # refine aesthetics
  scale_color_manual(values = pcols, drop = FALSE) +
  scale_fill_manual(values = pcols, drop = FALSE) +
  # add outcome labels
  geom_text(data = out_labs_print, 
            aes(label=outcome_lab, col = releff_cat, y = y_text), 
            x = 10.5, 
            hjust=0, # left align
            vjust = 1,
            size=2.5,
            inherit.aes = FALSE, show.legend = FALSE) +
  #
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  scale_x_continuous(breaks=seq(10,90,by=20)) +
  coord_cartesian(ylim=c(1,3.4)) +
  labs(x="Down sampled trial size: number of geographically proximate matched pairs",
       y="Relative efficiency of\ngeographic pair matching",
       ) +
  guides(colour = guide_legend(title = "Relative efficiency\nwith 90 pairs",nrow = 1)) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    # legend.text = element_text(size=12),
    legend.title = element_text(size=8),
    # axis.text = element_text(size=10),
    # axis.title = element_text(size=12),
    strip.text = element_text(size=10),
    plot.margin = margin(t = 14, r = 7, b = 7, l = 7, unit = "pt")
    )
plot_reff_b

```


# Kenya

## Correlation across scales
For each outcome, estimate the correlation between paired means.

Repeat estimates from 10 up to 72 pairs. 

For geographically proximate samples, compute estimates based on all 72 blocks selected as the initial sampling point, in turn (all permutations).


```{r kenya pairwise correlation by number of pairs}
#-------------------------------
# Kenya
#
# Geographically proximate 
# estimates across the 72
# possible permutations
#-------------------------------
start_time <- Sys.time()
cor_geo_k <- foreach(outi = levels(dcl_k$outcome_lab), .combine = rbind) %do% {
  foreach(npairs = c(seq(10,70,by=2),72), .combine = rbind) %:% 
  foreach(booti = 1:1000, .combine = rbind) %dopar% {
    
    # set seed within %dopar% for perfect reproducibility
    set.seed(booti)
    
    # randomly sample a block and select the n closest pairs
    # identify the geographically n proximate npairs to
    sampi <- geosample(distframe = distmat_k, nclosest = npairs)
    
    # then draw a bootstrap resample from the set of n closest pairs
    bsampi <- sample(1:npairs,size=npairs,replace=TRUE)
    bsamp  <- sampi$block_select[bsampi]
    bdist  <- sampi$block_dist[bsampi]
    
    di <- data.frame(block=as.character(bsamp), outcome_lab = outi) %>%
      left_join(dcl_k, by=c("outcome_lab","block"))
    
    # estimate outcome correlation, weighted by pair-wise N
    wY1bar <- di$Y1bar*di$w
    wY0bar <- di$Y0bar*di$w
    corri <- 2*cov(wY1bar,wY0bar) / ( var(wY1bar) + var(wY0bar) )
    
    # save the minimum and maximum distances between clusters
    mindisti <- min(bdist[as.numeric(bdist)>0])
    maxdisti <- max(bdist)
    
    # return results
    res <- data.frame(outcome_lab = outi, npairs, mindist = mindisti, maxdist = maxdisti, corr = corri)
    res
  }
}


cor_ests_k <- cor_geo_k %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl_k$outcome_lab)))

#----------------------------
# check time required to
# run bootstrap reps
#----------------------------
end_time <- Sys.time()
end_time - start_time
```


```{r kenya estimate mean correlation}
#-------------------------------
# estimate the mean and 95% CI
# for correlation across
# the different resampling
# replicates
#-------------------------------
sum_ests_k <- cor_ests_k %>% 
  group_by(outcome_lab,npairs) %>% 
  filter(!is.nan(corr)) %>%
  summarize(cor_est = mean(corr, na.rm=TRUE), 
            cor_min95 = quantile(corr,probs = 0.025, na.rm=TRUE),
            cor_max95 = quantile(corr,probs = 0.975, na.rm=TRUE),
            .groups = "keep") %>%
  mutate(x = npairs/2)


#-------------------------------
# estimate the average 
# maximum distance between
# clusters across
# the different resampling
# replicates
# this is the same for every
# outcome so can simplify
#-------------------------------
maxdist_ests_k <- cor_ests_k %>%
  # restrict to just one outcome (arbitrary)
  filter(outcome_lab == "Diarrhea") %>%
  group_by(npairs) %>% 
  summarize(maxdist_mu = mean(maxdist),
            maxdist_lb = quantile(maxdist,prob=0.025),
            maxdist_ub = quantile(maxdist,prob=0.975),
            .groups = "drop"
  ) %>%
  # convert distances to numeric and from m to km
  mutate(maxdist_mu = as.numeric(maxdist_mu)/1000,
         maxdist_lb = as.numeric(maxdist_lb)/1000,
         maxdist_ub = as.numeric(maxdist_ub)/1000
         )
```


## Estimate 95% CIs for relative efficiency

Use a bootstrap resampling approach, resampling pairs with replacement, to estimate the 95% confidence interval for relative efficacy of geographic pair matching with 79 matched pairs. This will enable us to get a sense for whether lower efficacy at smaller sample sizes is significantly lower, and not just a function of sampling variability.

```{r kenya bootstrap CIs for relative efficiency}
#-------------------------------
# group by outcome for bootstrap
#-------------------------------
dboot <- dcl_k %>%
  group_by(outcome_lab)

# estimate outcome correlation, weighted by pair-wise N
# (not used)
corx <- function(x,y,w) {
    xw <- x*w
    yw <- y*w
    wcorr <- 2*cov(xw,yw) / ( var(xw) + var(yw) )
    return(wcorr)
}

#-------------------------------
# boostrap resample within
# each outcome group
# 1000 iterations
#-------------------------------
cor_boot_k <- foreach(iteri = 1:1000, .combine = rbind) %dopar% {
  set.seed(iteri)
  di <- dboot %>% 
    slice_sample(prop = 1, replace = TRUE)
  corest <- di %>% group_map( ~ corx(x=.x$Y1bar, y=.x$Y0bar,w = .x$w))
  res <- data.frame(iter = iteri, outcome_lab = levels(di$outcome_lab), corr = unlist(corest))
  res
}

cor_boot_k2 <- cor_boot_k %>%
  mutate(outcome_lab = factor(outcome_lab, levels = levels(dcl_k$outcome_lab)))

#-------------------------------
# calculate percentile 95% CIs 
# from the bootstrap distribution
#-------------------------------
cor_boot_ci_k <- cor_boot_k2 %>%
  group_by(outcome_lab) %>%
  mutate(cor_min95 = quantile(corr, probs = 0.025),
         cor_max95 = quantile(corr, probs = 0.975)) %>%
  slice(1) %>%
  select(outcome_lab, cor_min95, cor_max95)

#-------------------------------
# transform correlation 95% CIs 
# into relative efficacy
#-------------------------------
releff_ci_k <- cor_boot_ci_k %>%
  mutate(releff_min95 = 1/(1-cor_min95),
         releff_max95 = 1/(1-cor_max95))

```

## Supplementary Figure 7 corr by size

Plot correlation and bootstrapped 95% CI by down-sampled trial size

```{r plot correlation by trial size kenya}

# calculate relative efficacy from correlation
dreleff_k <- sum_ests_k %>%
  mutate(releff = 1/(1-cor_est),
         releff_lb = 1/(1-cor_min95),
         releff_ub = 1/(1-cor_max95)) %>%
  # merge on the oucome groups 
  left_join( dcl_k %>% group_by(outcome_lab) %>% slice(1) %>% select(outcome_lab,outcome_group), by="outcome_lab") %>%
  # re-arrange order of the outcome groups for plot aesthetics
  mutate(outcome_lab_rev = factor(outcome_lab, levels = rev(levels(dcl_k$outcome_lab))) )

# make plot
pcols <- c(cbpal[2],cbpal[4],"gray50")
plot_corr_k <- ggplot(data=dreleff_k, aes(x = npairs, y = cor_est, color = outcome_group, fill = outcome_group)) +
  # facet over outcome
  facet_wrap(~outcome_lab_rev,ncol=4) +
  # plot results
  geom_hline(yintercept=0,color="gray40",lty="dashed") +
  geom_ribbon(aes(ymin = cor_min95, ymax = cor_max95), alpha=0.2, color = NA) +
  geom_line() +
  # refine aesthetics
  scale_color_manual(values = pcols) +
  scale_fill_manual(values = pcols) +
  # scale_color_brewer(palette = "Dark2") +
  # scale_fill_brewer(palette = "Dark2") +
 scale_y_continuous(breaks=seq(-0.6,1,by=0.2), labels=paste0(sprintf("%1.1f",seq(-0.6,1,by=0.2))) )+  
  scale_x_continuous(breaks=seq(10,90,by=20)) +
  labs(x="trial size: number of geographically proximate matched pairs",y="pair-level outcome correlation (r)") +
  guides(colour = guide_legend(title = "N pairs",nrow = 1)) +
  coord_cartesian(ylim=c(-0.6,1),xlim=c(10,72))+
  theme_minimal() +
  theme(
    legend.position = "none"  
    )
plot_corr_k


ggsave(here("output","washb-geopair-corr-geographic-scale-smallreps-kenya.png"), plot_corr_k, device = "png", width = 6, height = 8)
```

## Relative efficiency across geographic scales

```{r plot relative efficiency kenya compressed}
#-------------------------------
# calculate relative efficiency 
# from correlation
#-------------------------------
dreleff_k <- sum_ests_k %>%
  mutate(releff = 1/(1-cor_est),
         releff_lb = 1/(1-cor_min95),
         releff_ub = 1/(1-cor_max95)) %>%
  # merge on the outcome groups 
  left_join( dcl_k %>% group_by(outcome_lab) %>% slice(1) %>% select(outcome_lab,outcome_group), by="outcome_lab") %>%
  # re-arrange order of the outcome groups for plot aesthetics
  mutate(outcome_lab_rev = factor(outcome_lab, levels = rev(levels(dcl_k$outcome_lab))) )

#-------------------------------
# add correlation at 72 pairs
# to color each outcome in the
# figure
#-------------------------------
cor72ests_k <- dreleff_k %>%
  ungroup() %>%
  filter(npairs == 72) %>%
  select(outcome_group,outcome_lab, cor72 = cor_est, releff72 = releff) %>%
  # for purely asethetic visualization, bump waz up
  # a teeny tiny bit to be in the next cat 1.39 -> 1.4
  mutate(releff72 = ifelse(outcome_lab=="weight-for-age z",releff72+0.02,releff72)) %>%
  arrange(outcome_group, desc(releff72))

dreleff_k <- dreleff_k %>%
  left_join(cor72ests_k, by = c("outcome_group","outcome_lab"))

#-------------------------------
# store outcome labels to add to
# plot
#-------------------------------
out_labs_print_k <- cor72ests_k %>%
  group_by(outcome_group) %>%
  arrange(outcome_group, desc(releff72)) %>%
  mutate(outcome_order = seq_along(outcome_lab)) %>%
  mutate(y_text = 3.65 - outcome_order*0.15) %>% 
  ungroup() %>%
  mutate(releff_cat = case_when(
    releff72 <1.4 ~ "<1.4",
    releff72 >= 1.4 & releff72 < 1.8 ~ "1.4 to 1.8",
    releff72 >= 1.8 & releff72 < 2.2 ~ "1.8 to 2.2",
    releff72 >= 2.2 & releff72 < 2.6 ~ "2.2 to 2.6",
    releff72 >= 2.6 & releff72 < 3.0 ~ "2.6 to 3.0",
    releff72 >3.0 ~ ">3.0",
    TRUE ~ as.character(NA)
  ),
  releff_cat = factor(releff_cat, levels=c("<1.4","1.4 to 1.8","1.8 to 2.2","2.2 to 2.6","2.6 to 3.0",">3.0"))
         ) %>%
  select(outcome_group,outcome_lab,releff_cat, outcome_order, y_text)
  
dreleff_k <- dreleff_k %>%
  left_join(out_labs_print_k, by = c("outcome_group","outcome_lab"))

#-------------------------------
# make plot
#-------------------------------
# pcols <- rev(nytpal[2:7])
# pcols <- c("gray40",nytpal[c(7,6,4,3,2)])
pcols <- viridis(n=6,begin=0.1,end=0.8,option="D",direction=1)

plot_reff_k <- ggplot(data=dreleff_k, aes(x = npairs, y = releff, color = releff_cat, fill = releff_cat, group = outcome_lab)) +
  # facet over outcome
  facet_grid(~outcome_group) +
  # plot results
  # geom_ribbon(aes(ymin = releff_lb, ymax = releff_ub), alpha=0.2, color = NA, show.legend = FALSE) +
  geom_line() +
  # refine aesthetics
  scale_color_manual(values = pcols, drop = FALSE) +
  scale_fill_manual(values = pcols, drop = FALSE) +
  # add outcome labels
  geom_text(data = out_labs_print_k, 
            aes(label=outcome_lab, col = releff_cat, y = y_text), 
            x = 10.5, 
            hjust=0, # left align
            vjust = 1,
            size=2.5,
            inherit.aes = FALSE, show.legend = FALSE) +
  #
  scale_y_continuous(breaks=seq(1,3.4,by=0.4))+
  scale_x_continuous(breaks=seq(10,90,by=20)) +
  coord_cartesian(ylim=c(1,3.4), xlim=c(10,90)) +
  labs(x="Down sampled trial size: number of geographically proximate matched pairs",
       y="Relative efficiency of\ngeographic pair matching"
       ) +
  guides(colour = guide_legend(title = "Relative efficiency\nwith 72 pairs",nrow = 1)) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    # legend.text = element_text(size=12),
    legend.title = element_text(size=8),
    # axis.text = element_text(size=10),
    # axis.title = element_text(size=12),
    strip.text = element_text(size=10),
    plot.margin = margin(t = 14, r = 7, b = 7, l = 7, unit = "pt")
    )

plot_reff_k


```

# Figure 3 REFF by size
```{r combined plot overall}
#-------------------------------
# create an overall composite
# figure
#-------------------------------
reff_comp <- plot_grid(plot_reff_b,
                       plot_reff_k,
                       ncol=1,nrow=2,
                       rel_heights = c(1,1),
                         labels=c("a","b"),
                         label_size = 18,
                       hjust = 0
                         )

ggsave(here("output","washb-geopair-relative-efficiency-geographic-scale.pdf"), 
       reff_comp, 
       device = "pdf", 
       width = 180, height = 180, units = "mm")
```


# Supplementary Fig 8 max dist

Create a combined maximum distance figure that includes the two countries
```{r combined max distance fig}
#------------------------
# bind the max dist ests
# for the two countries
#------------------------
maxdist_ests <- maxdist_ests_b %>%
  mutate(country = "Bangladesh trial") %>%
  bind_rows(maxdist_ests_k) %>%
  mutate(country = ifelse(is.na(country),"Kenya trial",country))



#-------------------------------
# plot average max distance
# by number of pairs
#-------------------------------
plot_dist <- ggplot(data=maxdist_ests, aes(x=npairs, y=maxdist_mu, color=country, fill=country)) +
  geom_ribbon(aes(ymin=maxdist_lb,ymax=maxdist_ub),alpha=0.2, color=NA)+
  geom_line() +
  labs(x="Number of geographically proximate matched pairs",y="Average maximum distance between\ngeographically proximate matched pairs (km)") +
  scale_x_continuous(breaks = seq(10,90,by=10)) +
  scale_y_continuous(breaks = seq(0,120,by=20)) +
  scale_color_manual(values=cbpal[2:3]) +
  scale_fill_manual(values=cbpal[2:3]) +
  guides(color = guide_legend(title=""), fill = guide_legend(title="")) + 
  coord_cartesian(xlim=c(10,90),ylim=c(0,120)) +
  theme_minimal() +
  theme(
    # tweak the vertical margins to align with the Reff plot (above)
    legend.position = c(0.24,0.91),
    plot.margin = (unit(c(25, 5.5, 45, 5.5), "pt")),
    panel.grid.minor = element_blank(),
    axis.text = element_text(size=10),
    axis.title = element_text(size=12)
  )

plot_dist

ggsave(here("output","washb-geopair-maxdist-by-size.pdf"), 
       plot_dist, 
       device = "pdf", 
       width = 140, 
       height = 140,
       units = "mm")
```


# Save estimates
```{r save estimates}
#-------------------------------
# save the estimates for later
# reuse
#-------------------------------
dreleff <- dreleff_b %>%
  mutate(country="Bangladesh") %>%
  bind_rows(dreleff_k) %>%
  mutate(country = ifelse(is.na(country),"Kenya",country))

write_rds(dreleff, file = here("output","washb_releff_ests_by_number_of_pairs.rds"))
write_csv(dreleff, file = here("output","washb_releff_ests_by_number_of_pairs.csv"))


```

# Session Info
```{r session info}
sessionInfo()
```
