==> R/10-washb-geopair-pubmed-search-crt-fig.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "PubMed search results for CRTs"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: default
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

Create a chart from a PubMed search query `[("cluster randomized trial") OR ("group randomized trial")]`.  Search date: Sep 15, 2023

https://pubmed.ncbi.nlm.nih.gov/?term=%5B%28%22cluster+randomized+trial%22%29+OR+%28%22group+randomized+trial%22%29%5D+ 

Results downloaded to a `.csv` file, which is read-in below to chart

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))


#----------------------------
# source the shared functions file
#----------------------------
source(here("R","washb-geopair-Functions.R"))

#----------------------------
# Box data file paths
#----------------------------
Box_data_directory    <- "~/Library/CloudStorage/Box-Box/washb-geopair/data/"


```

# PubMed search

```{r pubmed search}
#-------------------------------
# load pubmed search 
#-------------------------------
pubmed_res <- read_csv(here(Box_data_directory,"untouched/PubMed_Timeline_Results_by_Year_[(\"cluster randomized trial\") OR (\"group randomized trial\")].csv"), skip = 1) 

as.data.frame(pubmed_res)
```

# Model fit for trend

```{r model fit for exponential trend}
#-------------------------------
# limit the data to 2000 to 2021
#-------------------------------
pubmed_res2 <- pubmed_res %>% 
  filter(Year>=2000 & Year <= 2022)

#-------------------------------
# fit an exponential model
# and get predicted values
#-------------------------------
pubmed_fit <- lm(log(Count) ~ Year, data = pubmed_res2)
summary(pubmed_fit)
pubmed_res2$pred <- exp(predict(pubmed_fit))

```

# Make figure

```{r make pubmed figure, fig.width=6}
plotyticks <- seq(0,500,by=100)
plot_pubmed_res <- ggplot(
  data = pubmed_res2,
  aes(x=Year)
  ) +
  geom_bar(aes(y=Count),stat = "identity", fill=cbpal[6],alpha=0.5, color = NA) + 
  geom_hline(yintercept = plotyticks[-1], color="white") +
  geom_line(aes(y=pred), color = "black") + 
  annotate("text",x=2018,y=450,label="exponential fit",hjust=1) +
  scale_y_continuous(breaks=plotyticks) +
  coord_cartesian(xlim=c(2000,2022),ylim=c(0,500)) +
  labs(x = "Year", y = "PubMed search results")+
  theme_classic()

plot_pubmed_res

ggsave(filename=here("output","washb-geopair-pubmed-search-results.pdf"),plot = plot_pubmed_res,device="pdf",width=4,height=4)
```

# Session Info
```{r session info}
sessionInfo()
```
