==> R/12-washb-geopair-spatial-heterogeneity.Rmd <==
---
title: "Geographic pair matching in large-scale cluster randomized trials"
subtitle: "Spatial heterogeneity in treatment effects"
author: "Ben Arnold ben.arnold@ucsf.edu"
date: "updated: `r Sys.time()`"
output: 
  html_document:
    theme: default
    highlight: default
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: true
      smooth_scroll: true
---

# Summary

This script conducts an analysis of an outcome in the Bangladesh trial (diarrhea) and an outcome in the Kenya trial (Ascaris sp. infection). 

These two outcomes were chosen as illustrative examples because both outcomes exhibit spatial heterogeneity in the control group means, yet provide contrasting examples of spatial heterogeneity in the average treatment effect (ATE) — some heterogeneity of the effect on diarrhea in Bangladesh, but almost none for Ascaris sp. in Kenya. 

For each outcome, we use a geostatistical regression model to krige (smooth) the mean in the control group between pair-matched locations, as well as the ATE.  As of Dec 2022, Wikipedia has one of the most concise summaries of the approach (https://en.wikipedia.org/wiki/Regression-kriging). We use a simple model that estimates the mean as a function of longitude and latitude plus a Matern spatial covariance function allowing for correlated values between locations. This is sometimes called _universal kriging_. 

Loosly following notation from Diggle and Giorgi (2016, https://doi.org/10.1080/01621459.2015.1123158) the model is:

$E[Y(x_i)] = d(x_i)^{'}\beta + S(x_i)$

where $d(x_i)$ includes an intercept plus longitude and latitude at each location, $x_i$, and $S(x_i)$ is a Gaussian process with mean zero and correlation function assumed to follow a Matern distribution between locations $x_i$, with $g(x) = g(x,x^{'}) = \left[S(x),S(x^')\right] = Matern(x; \nu,\rho)$ where $\nu$ is the smoothness parameter and $\rho$ estimates the decay in correlation with distance between locations.  We fit the model using maximum likelihood using the `spaMM` package in R.

Once fit, we then predict over a fine grid of the study area, trimmed to locations within study-region subdistricts that are within 10km of study clusters. We chose 10km as a reasonable distance to extrapolate from observed values. Outcomes in both examples demonstrated correlation within 10-20 km based on the fitted regression kriging models.  Finally, we simulate from the model fit of the fixed effect and spatial covariance matrix 1,000 realizations of the outcome at each location, and compute the proportion of those realizations where there is a treatment benefit (a reduction in prevalence of diarrhea or Ascaris sp. infection): $Pr(\hat{E}[Y(x_i)] < 0)$. This is an approximate posterior probability of a treatment benefit under an uninformative prior. It is useful because it conveys information about both the magnitude of the treatment effect and its uncertainty (Diggle and Giorgi 2016 https://doi.org/10.1080/01621459.2015.1123158, Fronterre et al. 2020 https://doi.org/10.1093/infdis/jiz554).

# Preamble

```{r preamble, message = FALSE}
library(here)
here()
#----------------------------
# source the config file
#----------------------------
source(here("R","washb-geopair-Config.R"))

#----------------------------
# source the shared functions file
#----------------------------
source(here("R","washb-geopair-Functions.R"))


#----------------------------
# Box data file paths
#----------------------------
Box_data_directory    <- "~/Library/CloudStorage/Box-Box/washb-geopair/data/final/"
Box_geodata_directory <- "~/Library/CloudStorage/Box-Box/washb-geopair/data/geodata/"
```

# Bangladesh diarrhea

## Data prep

```{r load bangladesh data}
#-------------------------------
# load the study GPS data
# data created by
# washb-geopair-gps-data-processing.R
#
# NOT PUBLICLY AVAILABLE
# NEED TO USE ACCURATE LON/LAT FOR MAPPING
#-------------------------------

# Cluster-level GPS locations (centroids of compounds in each cluster)
dgps <- read_rds(paste0(Box_data_directory,"bangl_analysis_gps.rds")) %>%
  mutate(block = as.character(block),
         blockf = factor(block))
  
#----------------------------------
# Load the Bangladesh
# administrative shapefile
#----------------------------------
admin_b <- st_read(paste0(Box_geodata_directory,"bgd_adm_bbs_20201113_SHP"),layer = "bgd_admbnda_adm3_bbs_20201113")
  
admin_b_study <- admin_b %>%
  # add in one more zilla that is encircled by study zillas
  # to allow for the 10km spatial predictions Fulbaria (BD456120)
  filter(ADM3_PCODE %in% c(dgps$ADM3_PCODE,"BD456120"))

#-------------------------------
# read-in block-level level means
# these were estimated in the file
# 02-washb-geopair-relative-efficiency.Rmd
#-------------------------------
dcl <- read_rds(here("data","washb_pair_level_summary_estimates.rds")) 
```

```{r bangladesh make spatial data}
#-------------------------------
# Create 2 GPS datasetes for 
# the spatial kriging
#
# dataset 1 has all control and
# nutrition cluster centroids,
# used to identify 10km buffer around
# the control + nutrition clusters (n=360)
#
# dataset 2 has block
# level centroids, computed
# as the median lon/lat across
# clusters in each block (n=90)
#-------------------------------
dbgps_cn <- dgps %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  dplyr::select(clusterid,tr,block,blockf,lon,lat)

dbgps_block <- dgps %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  ungroup() %>%
  dplyr::select(block,blockf,tr,lon,lat) %>%
  group_by(block,blockf) %>%
  summarize(lon = median(lon),
            lat = median(lat), 
            .groups = "keep") %>%
  mutate(country = "Bangladesh") 

# join the block-level GPS coordinates to block-level means
dbgps <- dbgps_block %>%
  left_join(dcl, by = c("country","block"))
```

```{r make bangladesh prediction grid}
#---------------------------
# identify the bounding box
# around subdistricts that
# contain the WASHB Bangladesh
# clusters
#---------------------------
b_boxgrid <- st_make_grid(admin_b_study,
                          n=c(100,100),
                          what = "centers"
                          )

#---------------------------
# create a 10 km buffer around
# clusters
# 10 km was chosen as a reasonably
# short distance over which to
# make spatial predictions
#---------------------------

# get cluster centroids
dbgps_sf_utm <- dbgps_cn %>%
  ungroup() %>%
  st_as_sf(coords = c("lon", "lat"), crs=4326) %>%
  dplyr::select(geometry) %>%
  # transform to UTM for km units
  st_transform("+proj=utm +zone=46 +datum=WGS84 +units=km")

# identify a 10km buffer around each cluster
cl_buff <- st_buffer(dbgps_sf_utm, dist=10) %>%
  # identify their union, then convert to a single polygon
  summarise(geometry = st_union(geometry)) %>%
  st_cast("POLYGON") %>%
  # set the CRS back to WGS84
  st_transform(crs=4326)

```


## Spatial predictions of diarrhea and ATE

```{r bangladesh diarrhea mean spatial model fit}
#-------------------------
# data frame for
# diarrhea spatial modeling
#-------------------------
bpairs_diar <- dbgps %>% filter(outcome_lab=="Diarrhea") %>% ungroup()

#-------------------------
# fit spatial model for
# diarrhea prevalence in 
# the control group
#-------------------------
(b_fitdiar <- spaMM::fitme(Y0bar ~  lat + lon + Matern(1|lat + lon),
                           data = bpairs_diar,
                           family = gaussian(link = "identity")) )
```
```{r bangladesh diar correlation by distance}
# (see function in washb-geopair-Functions.R)
(plot_materncor_diar <- plot_matern_corr(b_fitdiar) )
ggsave(filename=here("output","washb-geopair-bangladesh-diar-materncor.png"),plot = plot_materncor_diar,device="png",width=3.3,height=3.5)
```

Use the model to make spatial predictions over the study area grid.  

```{r bangladesh diar spatial model predictions}
#--------------------------------------
# get model predictions on a grid
# (see function in washb-geopair-Functions.R)
#--------------------------------------
b_preds_diar <- get_grid_preds(input_grid=b_boxgrid, 
                              spamm_model_fit = b_fitdiar)
#--------------------------------------
# convert predicted surface of the mean
# to a raster to make it a regular
# set of tiles (rather than points), 
# then to tibble for ggplot2
# scale prevalence by 100
# to be in units of percent (%)
#--------------------------------------
b_preds_diar_coords <- st_coordinates(b_preds_diar)
b_preds_diar_raster <- points_to_raster(
  x = b_preds_diar_coords[,1],
  y = b_preds_diar_coords[,2],
  z = b_preds_diar$pred*100,
  mask1 = admin_b_study,
  mask2 = cl_buff,
  crop1 = admin_b_study
)
b_preds_diar_raster2 <- raster_to_tibble(b_preds_diar_raster)
```


```{r map diar predicted from control}
map_diar <- ggplot(data=b_preds_diar_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_b_study,fill=NA) +
  coord_sf(crs = 4326) +
  scale_fill_viridis(option = "D", direction = 1,
                     limits = c(0,10),
                        na.value = NA,
                        breaks = seq(0,10,by=2),
                        guide=guide_colorbar(title="Diarrhea prevalence (%)", 
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "white",
                                           ticks.linewidth = 1,
                                           frame.colour = "white"
                        )
                        ) +
  geom_point(data=dbgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  # geom_point(data=dbgps_cn,aes(x=lon,y=lat),pch=16,color="gray40",size=0.8) +
  labs(x="Latitude",y="Longitude",tag="a") +
  annotation_scale(location = "br", width_hint = 0.2) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_diar
```

Use a geostatistical model to krige pair-level treatment effect estimates. Assume a Matern spatial covariance. Weight pairs by the number of measurements in each block.

```{r bangladesh diarrhea diff spatial model fit}
#-------------------------
# fit spatial model for
# differences
#-------------------------
(b_fitdiar_pairdiff <- spaMM::fitme(diff ~  lat + lon + Matern(1|lat + lon),
                           data = bpairs_diar,
                           family = gaussian(link = "identity")) )
```
Plot the estimated correlation function based on model fit parameters.

```{r bangladesh diar diff correlation by distance}
# (see function in washb-geopair-Functions.R)
(plot_materncor_diar_diff <- plot_matern_corr(b_fitdiar_pairdiff) )
ggsave(filename=here("output","washb-geopair-bangladesh-diar-diff-materncor.png"),plot = plot_materncor_diar_diff,device="png",width=3.3,height=3.5)
```

Use the model to make spatial predictions over the study area grid.  

```{r bangladesh diar diff spatial model predictions}
#--------------------------------------
# get model predictions on a grid
# (see function in washb-geopair-Functions.R)
#--------------------------------------
b_preds_diar_diff <- get_grid_preds(input_grid=b_boxgrid, 
                              spamm_model_fit = b_fitdiar_pairdiff)

#--------------------------------------
# convert predicted surface of mean
# diff to a raster to make it a regular
# set of tiles (rather than points), 
# then to tibble for ggplot2
# scale prevalence difference by 100
# to be in units of percent (%)
#--------------------------------------
b_preds_diar_diff_coords <- st_coordinates(b_preds_diar_diff)
b_preds_diar_diff_raster <- points_to_raster(
  x = b_preds_diar_diff_coords[,1],
  y = b_preds_diar_diff_coords[,2],
  z = b_preds_diar_diff$pred*100,
  mask1 = admin_b_study,
  mask2 = cl_buff,
  crop1 = admin_b_study
)
b_preds_diar_diff_raster2 <- raster_to_tibble(b_preds_diar_diff_raster)
```


```{r map diar predicted ATE from pairs}
PRGncols <- brewer.pal(n=11,name="PRGn")
map_diar_pairdiff <- ggplot(data=b_preds_diar_diff_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_b_study,fill=NA) +
  coord_sf(crs = 4326) +
  scale_fill_gradient2(low=PRGncols[10], mid="white", high=PRGncols[2],
                       limits = c(-8,2),
                        na.value = NA,
                        breaks = seq(-8,2,by=2),
                        guide=guide_colorbar(title="ATE: difference in diarrhea prev. (%)", 
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "black",
                                           ticks.linewidth = 1,
                                           frame.colour = "black"
                        )
                        ) +
  geom_point(data=dbgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  labs(x="Latitude",y="Longitude",tag="b") +
  annotation_scale(location = "br", width_hint = 0.2) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_diar_pairdiff
```

Use a parametric bootstrap simulation of the model parameters, conditional on the spatial model fit, to estimate an approximate posterior distribution of treatment effects at each location. 

```{r diar simulate posterior}
#--------------------------------------
# simulate from the conditional outcome
# model to get posterior probability of
# a positive benefit of intervention,
# ATE > 0, at each location
#--------------------------------------
#--------------------------------------
# create prediction grid dataset  
#--------------------------------------
b_boxgrid_coords <- st_coordinates(b_boxgrid)
b_predgrid <- b_boxgrid %>%
  as_tibble()
b_predgrid$lon <- as.double(b_boxgrid_coords[,1])
b_predgrid$lat <- as.double(b_boxgrid_coords[,2])

#--------------------------------------
# parametric bootstrap simulations
# of the posterior outcome distribution
# at each point in the grid
#--------------------------------------
sims_diar <- simulate(b_fitdiar_pairdiff,
                 type = "predVar",
                 variances=list(linPred=TRUE,disp=TRUE),
                 newdata = b_predgrid,
                 nsim = 1000,
                 seed = 124701
                 )

#--------------------------------------
# mean and 
# 95% CI of the parametric
# bootstrap
#--------------------------------------
sims_diar_mu <- apply(sims_diar, 1, FUN = function(x) quantile(x, prob=c(0.5)))
sims_diar_ci <- apply(sims_diar, 1, FUN = function(x) quantile(x, prob=c(0.025,0.975)))

#--------------------------------------
# posterior probability that the 
# difference in Diarrhea prev <0
#--------------------------------------
sims_diar_probeff <- apply(sims_diar, 1, FUN = function(x) mean(ifelse(x<0,1,0),na.rm=TRUE))

#--------------------------------------
# combine spatial predictions into a 
# data frame for mapping
#--------------------------------------
sims_diar_map <- b_predgrid %>%
  mutate(fit = sims_diar_mu,
         fit_lb = sims_diar_ci[1,],
         fit_ub = sims_diar_ci[2,],
         fit_ci_width=fit_ub-fit_lb,
         fit_probeff = sims_diar_probeff)
```

```{r diar posterior convert to raster}
#--------------------------------------
# convert simulation surfaces to a raster 
# to make it a regular set of tiles 
# (rather than points), 
# then to tibble for ggplot2
#--------------------------------------
diar_probeff_raster <- points_to_raster(
  x = b_predgrid$lon,
  y = b_predgrid$lat,
  z = sims_diar_map$fit_probeff,
  mask1 = admin_b_study,
  mask2 = cl_buff,
  crop1 = admin_b_study
)

diar_probeff_raster2 <- raster_to_tibble(diar_probeff_raster)
```

```{r map diar posterior}
map_diar_posterior <- ggplot(data=diar_probeff_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_b_study,fill=NA) +
  coord_sf(crs = 4326) +
  scale_fill_viridis(option="G", limits = c(0,1), direction = 1, breaks=seq(0,1,by=0.2), na.value = NA,
                      guide=guide_colorbar(title="Posterior Pr(diarrhea ATE < 0)",
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "white",
                                           ticks.linewidth = 1)
                      ) +
  geom_point(data=dbgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  labs(x="Latitude",y="Longitude",tag="c") +
  annotation_scale(location = "br", width_hint = 0.2) +
  annotation_north_arrow(location = "br", which_north = "true", 
        pad_x = unit(0.05, "in"), pad_y = unit(0.25, "in"),
        height=unit(0.75,"cm"), width=unit(0.75,"cm"),
        style = north_arrow_fancy_orienteering
        ) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_diar_posterior
```

# Kenya Ascaris

## Data prep

```{r load kenya data}
#-------------------------------
# load the study's GPS coordinates
# created by 
# washb-geopair-gps-data-processing.R
#
# NOT PUBLICLY AVAILABLE -- 
# NEED TO USE ACCURATE LON/LAT FOR MAPPING
#-------------------------------

# Cluster-level GPS locations (centroids of compounds in each cluster)
dgps_k <- read_rds(file=here(Box_data_directory,"kenya_analysis_gps.rds")) %>%
  # filter out blocks that were not complete for control,N, N+WSH
  filter(block %in% unique(dcl$block[dcl$country=="Kenya" & dcl$outcome_lab=="Ascaris sp."])) %>%
  mutate(block = as.character(block),
         blockf = factor(block)) 
  
#----------------------------------
# Load the Kenya
# administrative shapefile
#----------------------------------
admin_k <- st_read(paste0(Box_geodata_directory,"ken_adm_iebc_20191031_shp"),layer = "ken_admbnda_adm2_iebc_20191031")
  
# there are a small number of clusters (n=4) in Kenya 
# that seem to be geolocated just over the line 
# into a new sub-county. By hand, recode the 
# administrative information for these clusters 
# to be the same as those for the other clusters 
# in the same randomization block.
dgps_k <- dgps_k %>%
  mutate(
    
    ADM2_PCODE = ifelse(ADM2_PCODE=="KE037202" & block==3,"KE037205",ADM2_PCODE),
    ADM2_EN = ifelse(ADM2_PCODE=="KE037202" & block==3,"Mumias East",ADM2_EN),
    
    ADM2_PCODE = ifelse(ADM2_PCODE=="KE039220" & block==6,"KE039219",ADM2_PCODE),
    ADM2_EN = ifelse(ADM2_PCODE=="KE039220" & block==6,"Bumula",ADM2_EN),
    
    ADM2_PCODE = ifelse(ADM2_PCODE=="KE042239" & block==61,"KE038214",ADM2_PCODE),
    ADM2_EN = ifelse(ADM2_PCODE=="KE042239" & block==61,"Luanda",ADM2_EN),
    
    ADM2_PCODE = ifelse(ADM2_PCODE=="KE041235" & block==61,"KE038214",ADM2_PCODE),
    ADM2_EN = ifelse(ADM2_PCODE=="KE041235" & block==61,"Luanda",ADM2_EN)
  )

# limit sub counties to thos that overlap
# study block centroids
# also include Kanduyi (KE039220) because it
# is narrow and encircled by the study subcounties
admin_k_study <- admin_k %>%
  filter(ADM2_PCODE %in% c(dgps_k$ADM2_PCODE,"KE039220"))

```

```{r kenya make spatial data}
#-------------------------------
# Identify block centroids for
# the spatial kriging
#
# dataset 1 has all control and
# nutrition cluster centroids,
# used for getting a 10km buffer
# of the control + nutrition clusters (n=316)
#
# dataset 2 has block
# level centroids, computed
# as the median lon/lat across
# clusters in each block (n=79)
#-------------------------------
dkgps_cn <- dgps_k %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85)) %>%
  dplyr::select(clusterid,tr,block,blockf,lon,lat)

dkgps_block <- dgps_k %>%
  filter(tr %in% c("Control","Nutrition","Nutrition + WSH")) %>%
  filter(!block %in% c(11,13,20,23,30,33,37,40,49,51,54,56,65,72,80,84,85)) %>%
  ungroup() %>%
  dplyr::select(block,blockf,tr,lon,lat) %>%
  group_by(block,blockf) %>%
  summarize(lon = median(lon),
            lat = median(lat), 
            .groups = "keep") %>%
  mutate(country = "Kenya") 

# join the block-level GPS coordinates to block-level means
# (loaded above in the Bangladesh data prep section)
dkgps <- dkgps_block %>%
  left_join(dcl, by = c("country","block"))
```

```{r make kenya prediction grid}
#---------------------------
# identify the bounding box
# around subdistricts that
# contain the WASHB Bangladesh
# clusters
#---------------------------
k_boxgrid <- st_make_grid(admin_k_study,
                          n=c(100,100),
                          what = "centers"
                          )

#---------------------------
# create a 10 km buffer around
# clusters
# 10 km was chosen as a reasonably
# short distance over which to
# make spatial predictions
#---------------------------

# get cluster centroids
dkgps_sf_utm <- dkgps_cn %>%
  ungroup() %>%
  st_as_sf(coords = c("lon", "lat"), crs=4326) %>%
  dplyr::select(geometry) %>%
  # transform to UTM for km units
  st_transform("+proj=utm +zone=46 +datum=WGS84 +units=km")

# identify a 10km buffer around each cluster
kcl_buff <- st_buffer(dkgps_sf_utm, dist=10) %>%
  # identify their union, then convert to a single polygon
  summarise(geometry = st_union(geometry)) %>%
  st_cast("POLYGON") %>%
  # set the CRS back to WGS84
  st_transform(crs=4326)

```

## Spatial predictions of Ascaris infection and ATE

Use a geostatistical model to krige outcome prevalence and pair-level treatment effect estimates. Assume a Matern spatial covariance. Weight pairs by the number of measurements in each block.

```{r kenya ascaris spatial model fit}
#-------------------------
# data frame for
# Ascaris spatial modeling
#-------------------------
kpairs_al <- dkgps %>% filter(outcome_lab=="Ascaris sp.") %>% ungroup()

#-------------------------
# fit spatial model for
# pair-level
# differences in Ascaris infection
#-------------------------
# mean difference
(k_fital <- spaMM::fitme(Y0bar ~  lat + lon + Matern(1|lat + lon),
                           data = kpairs_al,
                           family = gaussian(link = "identity")) )

```

```{r kenya plot ascaris correlation by distance}
# (see function in washb-geopair-Functions.R)
(al_matern_cor_plot <- plot_matern_corr(k_fital))
ggsave(filename=here("output","washb-geopair-kenya-al-materncor.png"),plot = al_matern_cor_plot,device="png",width=3.3,height=3.5)

```

```{r kenya ascaris spatial krige predictions}
#--------------------------------------
# get model predictions on a grid
# (see function in washb-geopair-Functions.R)
#--------------------------------------
k_preds_al <- get_grid_preds(input_grid=k_boxgrid, 
                              spamm_model_fit = k_fital)

#--------------------------------------
# convert predicted surface of mean
# diff to a raster to make it a regular
# set of tiles (rather than points), 
# then to tibble for ggplot2
#
# multiple prevalence by 100 to scale
# units to percent (%)
#--------------------------------------
k_preds_al_coords <- st_coordinates(k_preds_al)
k_preds_al_raster <- points_to_raster(
  x = k_preds_al_coords[,1],
  y = k_preds_al_coords[,2],
  z = k_preds_al$pred*100,
  mask1 = admin_k_study,
  mask2 = kcl_buff,
  crop1 = admin_k_study
)

k_preds_al_raster2 <- raster_to_tibble(k_preds_al_raster)
```

```{r map kenya ascaris in control group}
#--------------------------------------
# map the spatially kriged Ascaris
# prevalence
#--------------------------------------
map_k_al <- ggplot(data=k_preds_al_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_k_study,fill=NA) +
  scale_fill_viridis(option="D", direction = 1,
                     limits = c(0,51),
                        na.value = NA,
                        breaks = seq(0,50,by=10),
                        guide=guide_colorbar(title="Ascaris sp. prevalence (%)", 
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "white",
                                           ticks.linewidth = 1,
                                           frame.colour = "white"
                        )
                        ) +
  geom_point(data=dkgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  scale_x_continuous(breaks=seq(34.4,35.0,by=0.2))+
  coord_sf(crs = 4326,xlim=c(34.3,35.0)) +
  labs(x="Latitude",y="Longitude",tag="d") +
  annotation_scale(location = "bl", width_hint = 0.3, pad_y=unit(0.1,"cm")) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_k_al
```


```{r kenya ascaris spatial model fit of ate}
#-------------------------
# fit spatial model for
# pair-level
# differences in Ascaris infection
#-------------------------
(k_fital_diff <- spaMM::fitme(diff ~  lat + lon + Matern(1|lat + lon),
                           data = kpairs_al,
                           family = gaussian(link = "identity")) )

```

```{r kenya plot ascaris diff correlation by distance}
# (see function in washb-geopair-Functions.R)
(al_diff_matern_cor_plot <- plot_matern_corr(k_fital_diff))
ggsave(filename=here("output","washb-geopair-kenya-al-diff-materncor.png"),plot = al_diff_matern_cor_plot,device="png",width=3.3,height=3.5)

```


```{r kenya ascaris spatial krige diff predictions}
#--------------------------------------
# get model predictions on a grid
# (see function in washb-geopair-Functions.R)
#--------------------------------------
k_preds_al_diff <- get_grid_preds(input_grid=k_boxgrid, 
                              spamm_model_fit = k_fital_diff)

#--------------------------------------
# convert predicted surface of mean
# diff to a raster to make it a regular
# set of tiles (rather than points), 
# then to tibble for ggplot2
#
# multiple prevalence by 100 to scale
# units to percent (%)
#--------------------------------------
k_preds_al_diff_coords <- st_coordinates(k_preds_al_diff)
k_preds_al_diff_raster <- points_to_raster(
  x = k_preds_al_diff_coords[,1],
  y = k_preds_al_diff_coords[,2],
  z = k_preds_al_diff$pred*100,
  mask1 = admin_k_study,
  mask2 = kcl_buff,
  crop1 = admin_k_study
)

k_preds_al_diff_raster2 <- raster_to_tibble(k_preds_al_diff_raster)

```

```{r map kenya ascaris kriged ATE from pairs}
#--------------------------------------
# map the spatially kriged treatment
# effect for Ascaris
#--------------------------------------
PRGncols <- brewer.pal(n=11,name="PRGn")
map_k_al_pairdiff <- ggplot(data=k_preds_al_diff_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_k_study,fill=NA) +
  scale_fill_gradient2(low=PRGncols[10], mid="white", high=PRGncols[2],limits = c(-8,2),
                        na.value = NA,
                        breaks = seq(-8,2,by=2),
                        guide=guide_colorbar(title="ATE: difference in Ascaris sp. prev. (%)",
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "black",
                                           ticks.linewidth = 1,
                                           frame.colour = "black"
                        )
                        ) +
  geom_point(data=dkgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  scale_x_continuous(breaks=seq(34.4,35.0,by=0.2))+
  coord_sf(crs = 4326,xlim=c(34.3,35.0)) +
  labs(x="Latitude",y="Longitude",tag="e") +
  annotation_scale(location = "bl", width_hint = 0.3, pad_y=unit(0.1,"cm")) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_k_al_pairdiff

```


```{r kenya ascaris simulate posterior}
#--------------------------------------
# simulate from the conditional outcome
# model to get posterior probability of
# a positive benefit of intervention,
# ATE < 0, at each location
#--------------------------------------
#--------------------------------------
# create prediction grid dataset  
#--------------------------------------
k_boxgrid_coords <- st_coordinates(k_boxgrid)

k_predgrid <- k_boxgrid %>%
  as_tibble()
k_predgrid$lon <- as.double(k_boxgrid_coords[,1])
k_predgrid$lat <- as.double(k_boxgrid_coords[,2])

#--------------------------------------
# parametric bootstrap simulations
# of the posterior outcome distribution
# at each point in the grid
#--------------------------------------
sims_al <- simulate(k_fital_diff,
                 type = "predVar",
                 variances=list(linPred=TRUE,disp=TRUE),
                 newdata = k_predgrid,
                 nsim = 1000,
                 seed = 213123
                 )

#--------------------------------------
# mean and 
# 95% CI of the parametric
# bootstrap
#--------------------------------------
sims_al_mu <- apply(sims_al, 1, FUN = function(x) quantile(x, prob=c(0.5)))
sims_al_ci <- apply(sims_al, 1, FUN = function(x) quantile(x, prob=c(0.025,0.975)))

#--------------------------------------
# posterior probability that the 
# difference in Diarrhea prev <0
#--------------------------------------
sims_al_probeff <- apply(sims_al, 1, FUN = function(x) mean(ifelse(x<0,1,0),na.rm=TRUE))

#--------------------------------------
# combine spatial predictions into a 
# data frame for mapping
#--------------------------------------
sims_al_map <- k_predgrid %>%
  mutate(fit = sims_al_mu,
         fit_lb = sims_al_ci[1,],
         fit_ub = sims_al_ci[2,],
         fit_ci_width=fit_ub-fit_lb,
         fit_probeff = sims_al_probeff)
```

```{r ascaris posterior convert to raster}
#--------------------------------------
# convert simulation surfaces to a raster 
# to make it a regular set of tiles 
# (rather than points), 
# then to tibble for ggplot2
#--------------------------------------
al_probeff_raster <- points_to_raster(
  x = k_predgrid$lon,
  y = k_predgrid$lat,
  z = sims_al_map$fit_probeff,
  mask1 = admin_k_study,
  mask2 = kcl_buff,
  crop1 = admin_k_study
)

al_probeff_raster2 <- raster_to_tibble(al_probeff_raster)
```

```{r map kenya ascaris posterior}
#--------------------------------------
# map the probability of nutrition 
# benefit on Ascaris sp. infection
#--------------------------------------
map_k_al_posterior <- ggplot(data=al_probeff_raster2) + 
  geom_tile(aes(x=x,y=y,fill=value), na.rm=TRUE) +
  geom_sf(data=admin_k_study,fill=NA) +
  # continuous scale
  scale_fill_viridis(option="G", limits = c(0,1), direction = 1, breaks=seq(0,1,by=0.2), na.value = NA,
                      guide=guide_colorbar(title="Posterior Pr(Ascaris ATE < 0)",
                                           title.position = "top",
                                           title.hjust = 0.5,
                                           direction="horizontal",
                                           label.position = "bottom",
                                           label.hjust = 0.5,
                                           barheight=unit(10,"pt"),
                                           barwidth=unit(200,"pt"),
                                           ticks.colour = "white",
                                           ticks.linewidth = 1)
                      ) +
  geom_point(data=dkgps_block, aes(x=lon,y=lat),pch=21,size=2,color="black",fill=NA) +
  labs(x="Latitude",y="Longitude",tag="f") +
  scale_x_continuous(breaks=seq(34.4,35.0,by=0.2))+
  coord_sf(crs = 4326,xlim=c(34.3,35.0)) +
  annotation_scale(location = "bl", width_hint = 0.3, pad_y=unit(0.1,"cm")) +
  annotation_north_arrow(location = "bl", which_north = "true", 
        pad_x = unit(0.05, "in"), pad_y = unit(0.25, "in"),
        height=unit(0.75,"cm"), width=unit(0.75,"cm"),
        style = north_arrow_fancy_orienteering
        ) +
  theme_minimal() +
  theme(
    legend.position = "bottom",
    plot.tag = element_text(face="bold",size=16)
  )

map_k_al_posterior
```

# Create a composite figure

Include one outcome from each country: Diarrhea in Bangladesh and Ascaris in Kenya. They provide contrasting examples of no spatial heterogeneity (Ascaris) and some spatial heterogeneity (diarrhea) in the ATE, despite both having underlying spatial heterogeneity in the prevalence of the outcome.

```{r kriged ATE composite map, fig.width=8}
#--------------------------------------
# create a 2 x 3 composite figure a-f
#--------------------------------------
ate_map_composite <- grid.arrange(map_diar,
                                  map_diar_pairdiff,
                                  map_diar_posterior,
                                  map_k_al,
                                  map_k_al_pairdiff,
                                  map_k_al_posterior,
                                  nrow=2,ncol=3)
ggsave(filename=here("output","washb-geopair-ate-krige-maps.png"),plot = ate_map_composite,device="png",width=12,height=10)
```


# Plot Matern functions for all outcomes

Estimate and plot the Matern correlation function for all outcomes

## Bangladesh

```{r est matern for all outcomes bangladesh}
#-------------------------
# fit generalized kriege
# for each outcome and 
# store fits as a list
#-------------------------
outcomes <- unique(dbgps$outcome_lab)
boutfits <- foreach(outi = outcomes) %do% {
  di <- dbgps %>% filter(outcome_lab == outi) %>% ungroup()
  fiti <- spaMM::fitme(Y0bar ~  lat + lon + Matern(1|lat + lon),
                       data = di,
                       family = gaussian(link = "identity"))
  return(fiti)
}

```

```{r plot matern functions bangladesh}
#-----------------------------------
# note: the plot_matern_corr()
# function is defined in
# washb-geopair-Functions.R
#-----------------------------------
# plot kriege fits for each outcome
outcomes <- unique(dbgps$outcome_lab)
plot_materns_b <- foreach(outi = 1:length(outcomes)) %do% {
  ploti <- plot_matern_corr(boutfits[[outi]]) +
    # modify slightly from original fn for this 
    # small replicates version of the plot
    geom_line(aes(x=pdistkm, y = materncor)) +
    labs(title = paste0(outcomes[outi])) +
    theme(title = element_text(hjust=0.5,face="bold", size=14),
          axis.title = element_text(face="plain"))
  
}

# create composite from the stored list
plot_materns_b_comp <- grid.arrange(
  plot_materns_b[[1]], 
  plot_materns_b[[2]],
  plot_materns_b[[3]],
  plot_materns_b[[4]],
  plot_materns_b[[5]],
  plot_materns_b[[6]],
  plot_materns_b[[7]],
  plot_materns_b[[8]],
  plot_materns_b[[9]],
  plot_materns_b[[10]],
  plot_materns_b[[11]],
  plot_materns_b[[12]],
  plot_materns_b[[13]],
  plot_materns_b[[14]],
  nrow=5,ncol=3
)

ggsave(filename=here("output","washb-geopair-matern-corrs-bangladesh.png"),plot = plot_materns_b_comp,device="png",width=12,height=20)
```


## Kenya

```{r est matern for all outcomes kenya}
#-------------------------
# fit generalized kriege
# for each outcome and 
# store fits as a list
#-------------------------
outcomes <- unique(dkgps$outcome_lab)
koutfits <- foreach(outi = outcomes) %do% {
  di <- dkgps %>% filter(outcome_lab == outi) %>% ungroup()
  fiti <- spaMM::fitme(Y0bar ~  lat + lon + Matern(1|lat + lon),
                       data = di,
                       family = gaussian(link = "identity"))
  return(fiti)
}

```

```{r plot matern functions kenya}
#-----------------------------------
# note: the plot_matern_corr()
# function is defined in
# washb-geopair-Functions.R
#-----------------------------------
# plot kriege fits for each outcome
plot_materns_k <- foreach(outi = 1:length(outcomes)) %do% {
  ploti <- plot_matern_corr(koutfits[[outi]]) +
    # modify slightly from original fn for this 
    # small replicates version of the plot
    geom_line(aes(x=pdistkm, y = materncor)) +
    labs(title = paste0(outcomes[outi])) +
    theme(title = element_text(hjust=0.5,face="bold", size=14),
          axis.title = element_text(face="plain"))
  
}

# create composite from the stored list
plot_materns_k_comp <- grid.arrange(
  plot_materns_k[[1]], 
  plot_materns_k[[2]],
  plot_materns_k[[3]],
  plot_materns_k[[4]],
  plot_materns_k[[5]],
  plot_materns_k[[6]],
  plot_materns_k[[7]],
  plot_materns_k[[8]],
  plot_materns_k[[9]],
  plot_materns_k[[10]],
  plot_materns_k[[11]],
  plot_materns_k[[12]],
  nrow=4,ncol=3
)

ggsave(filename=here("output","washb-geopair-matern-corrs-kenya.png"),plot = plot_materns_k_comp,device="png",width=12,height=16)
```

# Session Info
```{r session info}
sessionInfo()
```
